==> 2_Building_InputTraits_Pts.Rmd <==
---
title: "Building Input Dataframes for MNMs"
date: "11/15/2022"
output: pdf_document
---

```{r setup, include=FALSE}
library(tidyverse); library(knitr); library(ggrepel); library(cowplot); library(MetBrewer)
library(Rphylopars); library(ggtree); library(treeplyr); library(phytools)
library(sf); library(geodata);library(stars)
library(nlme)
knitr::opts_chunk$set(echo = FALSE)
# species with robust trait availability and compared in MNM paper
mnm_spp<- c('Lithobates catesbeianus', 'Lithobates clamitans', 'Anaxyrus americanus', 'Lithobates sylvaticus', 'Pseudacris crucifer', 'Dryophytes versicolor', 'Acris blanchardi', 'Dryophytes chrysoscelis', 'Lithobates palustris', 'Anaxyrus fowleri', 'Scaphiopus holbrookii', 'Gastrophryne carolinensis', 'Lithobates areolatus')
# species identified as interesting for the comparision analysis
com_spp<- c( "Acris gryllus", "Acris crepitans",
               "Anaxyrus terrestris", "Anaxyrus fowleri", "Anaxyrus quercicus", 
               "Dryophytes avivoca",  "Dryophytes andersonii", "Dryophytes cinereus", 
               "Dryophytes chrysoscelis", "Dryophytes femoralis", "Dryophytes gratiosus",
               "Dryophytes squirellus", "Gastrophryne carolinensis", "Lithobates sevosus",
               "Lithobates capito", "Lithobates grylio", "Lithobates okaloosae", 
               "Lithobates heckscheri", "Lithobates sphenocephalus", "Lithobates areolatus", 
               "Lithobates virgatipes", "Pseudacris ornata", "Pseudacris feriarum", 
             "Pseudacris brachyphona", "Pseudacris fouquettei","Scaphiopus holbrookii")
focal_spp<-unique( c(mnm_spp, com_spp))
```

\tableofcontents

# Trait Imputation Methods

[Penone et al. 2014](https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.12232) reviewed imputation methods and [Johnson et al. 2021](https://onlinelibrary.wiley.com/doi/full/10.1111/geb.13185) addressed how bias affects the top methods. Generally, using phylogentic information reduces error but bias in trait completeness leads to different methods being *best*.

#### Bias in our trait dataset

In 0_Species_Selection.Rmd, we identified species that have trait data for atleast 3 of the following traits: CTmax, CTmin, Tpref, and Tmerge. This Rmd also identified traits for the species within our species list for the climate change vulnerability comparison paper. 

```{r atraiu2, echo=F, warning=F, message=F}
trait_db<-read_csv('ATraiU 2.0/ATraiU2_summary_values_2022SEPT.csv', show_col_types = F)
knitr::include_graphics("C:/Users/Owner/Desktop/Git/AnuranMNMs/ATraiU 2.0/manuscript/summary_figure_combo.jpg")
```

```{r trait completeness, echo=F}
kable(trait_db %>% 
        dplyr::select(-refIDs, -arrhythmic, -crepuscular, -nocturnal, -diurnal) %>%
        pivot_longer(-species) %>%
        group_by(species) %>% 
        filter(species %in% focal_spp, name !='Tforage_optim')%>% 
        summarize(ntraits=sum(is.na(value)), rel.n.traits=round(ntraits/7,3)*100),
caption='Missing Trait Data among focal species',
col.names = c('Species', 'N missing Traits', 'Missing Data %'))
```

To improve the fit of this imputation, for our focal species, we are recalculating means that remove traits from far localities (e.g., Canada). 

```{r}
# pull in the USGS data release full database
new_trait_mean<-#read.csv('G:/Shared drives/GCC Vulnerability Scaling Project/3_Physiological Trait Database/Data Release/ATraiU2_full_2022SEPT.csv') %>% 
  read.csv('C:/Users/Owner/Downloads/TRAD_full_2022SEPT.csv') %>%
  filter(species %in% focal_spp, #only species in our species list
         lifeStage %in% c('adult','unknown'), # only adults
         traitName != 'Activity', # activity is categorical and can't be inputed using the method we want
         !is.na(verbatimLocality)) %>% # needs to have a known location to be excluded
  # count how many trait values are available per species & trait type
  group_by(species, traitName) %>% mutate(n=n()) %>% 
  filter(n != 1) # if only one value, won't exclude
#new_trait_mean$verbatimLocality %>% unique()

trait_filt <- new_trait_mean %>% 
  # I looked though the traits by hand and removed those I thought too far away
  filter(!(verbatimLocality %in% c("James River and Big Sioux River, South Dakota",
                                 "44.57229°N, 97.02486°W", 
                                 "45°34'57\"N 64°58'37\"W",
                                 "44°33'N, 78°33'W", "45°35'N, 78°30'W",
                                 "44°50'N, 68°35'W", 
                                 "Meadow Valley Wash between Carp & Elgin, Lincoln County, Nevada",
                                 "Andes of Veneszuela",
                                 "NAD 27: UTM Zone 17: N5127200m, E304000m")), 
         !grepl('rhode island', verbatimLocality),
         !grepl('Rhode Island', verbatimLocality),
          !grepl('Northeast US', verbatimLocality),
         !grepl('Massachusetts', verbatimLocality),
         !grepl('Connecticut', verbatimLocality),
          !grepl("48 N, 123 W", verbatimLocality),
          !grepl("South Dakota", verbatimLocality),
           !grepl("northwest", verbatimLocality),
           !grepl("Maine", verbatimLocality),
           !grepl("Minnesota", verbatimLocality),
           !grepl("MN", verbatimLocality),
           !grepl("Ontario", verbatimLocality),
           !grepl("Quebec", verbatimLocality),
         !grepl("Ottawa", verbatimLocality),
         !grepl("Vancouver Island", verbatimLocality),
           !grepl("British Columbia", verbatimLocality))
# save the traits that are used to recalculate interspecific level traits
write.csv(trait_filt, 'intermediate_results/regional_interspecific_traits.csv', row.names = F)
```

```{r all trait location plot, echo=F}
# usa map
usa<-st_as_sf(map('state', fill=T, plot=F)) %>% st_make_valid() %>%
  st_transform(st_crs("+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=km +no_defs"))
# original spatial extent of what we consider the se
se_orig<-st_as_sf(map('state',c('mississippi', 'georgia', 'florida', 'south carolina', 
                           'north carolina', 'virginia', 'alabama', 'tennessee'), 
                 fill=T, plot=F)) %>% 
  st_make_valid() %>%
  st_transform(st_crs("+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=km +no_defs"))
# new spatial extent based on ecological boundaries
se<-read_sf('C:/Users/Owner/Documents/GISfile/WBD_National_GDB/WBD_National_GDB.gdb',layer='WBDHU2') %>%
  filter(huc2 %in% c('03','06','08'))

# fixing trait locations so that they will 
# join with the se object through the ID column and can be plotted 
trait_filtL<- trait_filt %>%
  mutate(ID=case_when(grepl('Texas', verbatimLocality)|grepl('TX', verbatimLocality) ~ 'texas',
                      grepl('Missouri', verbatimLocality) ~ 'missouri',
                      grepl('Illinois', verbatimLocality) | 
                        grepl('IL', verbatimLocality) ~ 'illinois',
                      grepl('Ohio', verbatimLocality) |
                        grepl(', OH', verbatimLocality) ~ 'ohio',
                      grepl('New Jersey', verbatimLocality) ~ 'new Jersey',
                      grepl('Kentucky',verbatimLocality) ~ 'kentucky',
                      grepl('Michigan', verbatimLocality)~'michigan', 
                      grepl('New York', verbatimLocality) |
                        grepl('NY', verbatimLocality) ~ 'new york',
                      grepl('Indiana', verbatimLocality)~'indiana',
                      grepl('Pennsylvania', verbatimLocality)~'pennsylvania',
                      grepl('Arkansas', verbatimLocality)~'arkansas',
                      grepl('Louisiana', verbatimLocality)~'louisiana',
                      grepl('Maryland', verbatimLocality)~'maryland',
                      grepl('Delaware', verbatimLocality)~'delaware',
                      grepl('rhode island', verbatimLocality) ~ 'rhode island',
                      grepl('Kansas', verbatimLocality) ~ 'kansas',
                      grepl('Oklahoma', verbatimLocality)|
                        verbatimLocality == 'UTM zone 15; 233905E, 4047104N, NAD 83' ~ 'oklahoma',
                      verbatimLocality %in% c("41 17' N, 83 37' W", "41 17' N, 83 13' W",
                                              "40 37' N, 83 17' W", "41° 17' N, 83° 37' W",
                                              "41° 17' N, 83° 13' W", "40° 37' N, 83° 17' W") ~ 'ohio',
                      verbatimLocality == "44°28'45\"N, 87°37'15\" W" ~ 'wisconsin',
                      verbatimLocality == "N 0695928 and W 4482919, zone 15 using NAD27 CONUS map datum in UTM" ~ 'illinois',
                      grepl('Alabama', verbatimLocality)~'alabama',
                         grepl('Florida', verbatimLocality)|
                           grepl('FL', verbatimLocality) ~ 'florida',
                         grepl('Georgia', verbatimLocality) | 
                           grepl('GA', verbatimLocality) ~ 'georgia',
                         grepl('Mississippi', verbatimLocality) ~'mississippi',
                         grepl('North Carolina', verbatimLocality) ~'north carolina',
                         grepl('South Carolina', verbatimLocality) ~ 'south carolina', 
                         grepl('Tennessee', verbatimLocality) ~ 'tennessee',
                         grepl('Virginia', verbatimLocality) ~ 'virginia',
         T ~ 'study area')) %>%
  filter(species %in% mnm_spp)
# identify states that had traits that were removed from our trait value pool
usa_traits<-left_join(usa, trait_filtL %>%
                        distinct(traitName, ID) %>%
                        group_by(ID) %>% summarize(n_traits=length(unique(traitName))),
  by='ID') %>%
  mutate(n_traits = factor(ifelse(ID %in% c('south dakota', 'nevada', "maine", "minnesota",
                                            "massachusetts", "connecticut","rhode island"), 
                           0, n_traits))) 
sf::sf_use_s2(FALSE)
usa_traits %>%
  ggplot()+
  geom_sf(aes(fill=n_traits))+
  geom_sf(data=se %>% st_make_valid() %>% st_union() %>% st_transform(st_crs(usa)) , fill=NA, size=1.5, color='black')+
  scale_fill_manual('N traits\nused', values=c(met.brewer('Benedictus', 10)[c(1, 6:10)]))+
  theme_void()
ggsave('results/FigS2_1_traits_removed.jpg', width=4, height=2.75)
sf::sf_use_s2(TRUE)
```

In the first pass, I decided the state-based cutoff but wanted to quantify how consistent this was across regions. I consider my edge states as Mississippi, Tennessee, and Virginia and calculated the diestance between the centroids of those edge states and the states that had trait values I kept. I found I shouldn't include some of the Northeast states I included, and adjusted the list above. Excluded states and their summed distance from our edge states' centroids.

```{r}
edge_sts<-usa_traits %>% filter(ID %in% c("mississippi", "tennessee", "virginia")) %>% 
  st_centroid() 

dist.mat<-usa_traits %>% filter(!is.na(n_traits),
                      !(ID %in% se$ID)) %>%
  st_distance(edge_sts)
colnames(dist.mat)<-edge_sts$ID
rownames(dist.mat)<-usa_traits %>% filter(!is.na(n_traits), !(ID %in% se$ID)) %>% pull(ID)
rowMeans(dist.mat)[rowMeans(dist.mat)>1000]
```

## Identify which species to use to investigate intraspecific trait variation

Using the code below, we identified Tpref as a way to investigate intraspecific trait variation's effect on MNMs. Both Tmerge, Tpref and CTmax had some multiple recorded values for species in our species list. Of the traits we varied in our mechanistic niche model, Peterson and Gade (2017) found that optimal temperature had the largest effect on MNM outputs and it does not directly affect the interpretation of warming tolerance. Thus, we chose four species to vary Tpref (values observed from the literature) in our MNM models. 

```{r explore multiple trait avail}
trait_filt %>% filter(traitName %in% c('CTmax','Tpref','Tmerge')) %>%
  mutate(n=n()) %>% arrange(desc(n)) %>%
  select(species, traitName, traitValue, verbatimLocality, traitSource) %>% 
  arrange(traitName, species) %>% View()
trait_filt %>% filter(traitName %in% c('CTmax','Tpref','Tmerge'), species %in% mnm_spp) %>%
  count() %>% pivot_wider(names_from=traitName, values_from=n) %>%
  mutate(ntotal=sum(c(Tmerge, Tpref, CTmax), na.rm=T))
trait_filt %>% filter(traitName %in% c('CTmax','Tpref','Tmerge'),
                      species %in% c('Acris blanchardi','Anaxyrus americanus',
                                     'Pseudacris crucifer', 'Lithobates clamitans')) %>% 
  select(species, traitName, traitValue, verbatimLocality, traitSource) %>% 
  arrange(species, verbatimLocality) %>% View()
```

## Outputting interspecific trait values 

```{r combine trait tables, warning=F, message=F}
# recalculate the trait value to use in our models 
# based only on trait values closer to the SE
trait_restrict <- trait_filt %>%  group_by(species, traitName) %>%
  dplyr::summarize(mean_val=case_when(traitName == 'CTmax' ~ max(as.numeric(traitValue)),
                               traitName == 'CTmin' ~ min(as.numeric(traitValue)),
                               T ~ mean(as.numeric(traitValue))), 
            sd_val=sd(as.numeric(traitValue)), .groups='drop') %>%
  distinct(species, traitName, mean_val, sd_val)
# join the ATraiU-TaDs summarized trait values to the local summarized trait values
both_trait_types<-trait_db %>% select(-refIDs) %>%
  pivot_longer(-species, names_to = "traitName") %>%
  left_join(trait_restrict, by = c("species", "traitName")) %>%
# upon inspection of G. carolinensis's high Tmerge, we should use min Tmerge
  mutate(value=ifelse(grepl('carolin', species) & traitName == 'Tmerge', 22.0, value))

# replace ATraiU-TaDs summarized trait values if local trait values available
trait_used<-both_trait_types %>%
  mutate(trait_value=ifelse(is.na(mean_val), value, round(mean_val,3))) %>%
  # change datafraome to wide instead of long
  select(species, traitName, trait_value) %>% 
  pivot_wider(names_from=traitName, values_from=trait_value)
```
```{r kept trait table, echo=F}
kable(both_trait_types %>%
  filter(!(traitName %in% c('nocturnal', 'diurnal', 'crepuscular', 'arrhythmic')),
         species %in% focal_spp) %>%
  mutate(`Species List`=case_when(species %in% mnm_spp & species %in% com_spp ~ 'both papers',
                            species %in% mnm_spp ~ 'mnm paper',
                            species %in% com_spp ~ 'comparison paper'),
         keptint=case_when(is.na(mean_val)& !is.na(value)~'ATraiU-TaDS', 
                           !is.na(mean_val) ~ 'SE recalc',
                           T ~ 'missing')) %>% 
    count(`Species List`,keptint) %>% 
    pivot_wider(names_from = keptint, values_from=n),
  caption='Types of trait values used as inputs to our MNM models')
```
```{r compare trait values, fig.width=6.5, fig.height=3, echo=F}
both_trait_types %>%
  filter(!is.na(mean_val),
         traitName != 'Tbask', !grepl('Tforage', traitName))  %>%
  mutate(trait_dif=value-mean_val) %>%
  ggplot()+
  geom_point(aes(y=species, x=trait_dif))+
  facet_wrap(~traitName, scales='free', ncol=3)+
  scale_x_continuous('Trait Difference\nEntire Range Trait Value - Within 1000km of the SE Trait Value')+
  theme_classic()+
  theme(strip.background=element_blank(),
        axis.text.y = element_text(face='italic'),
        axis.text.x=element_text(size=8),
        axis.title.y=element_blank())
```

Based on this quick comparison, we can see expected differences between interspecific trait values calculated for the species entire range and interspecific trait values calculated just using trait values measured within a mean of 1000km of the edge our study region. An undexpected difference is mean CTmax for L. catesbeianus is higher when calculated when using the entire range, despite the SE US region having hotter summers. Mean CTmin is higher in the SE, likely because of its milder winters. Mass is mostly consistent except for much (~25%) larger L. catesbeianus (American bullfrog). Preferred temperatures are higher when we restrict values to the SE and emergence temperatures are later, both of which might be expected because of hotter temperatures and more resources (thus less need to compete for resources by emerging sooner).   

So when multiple traits are available for our focal species, we only use traits within a mean of 1000km of the edge states of our study region. When only one trait is available for a species and trait, we use that trait. We did not restrict the trait values for non-focal SE species (which are used in the phylogenetic imputation process) because those species might have large ranges and are already likely trait limited. Below is some code to describe the amount of traits removed from ATraiU-TaDS.

```{r trait removal results code}
 # number of traits removed
nrow(new_trait_mean %>% filter(species %in% mnm_spp)) - 
  nrow(trait_filt %>% filter(species %in% mnm_spp))
# number of species with count of traits removed
anti_join(new_trait_mean, trait_filt,
          by = c("species", "traitName", "lifeStage", "sex", "traitValue", "traitUnit",
"traitVariation", "traitVariationType", "traitAcclimation", "verbatimLocality", "traitSource",
"Notes", "traitEvidenceCategory", "canonicalName", "scientificName", "family", "genus",
"usageKey", "refID", "outlier", "itisTSN", "n")) %>% ungroup() %>%
  filter(species %in% mnm_spp) %>% count(species)
trait_filt %>% ungroup() %>% 
  filter(species %in% mnm_spp) %>% count(species) %>%
  # average trait values that went into the recalculated number
  summarize(mean(n), sd(n), sum(n))
```


## Anuran Phylogeny

Now that we have interspecific traits that are closer to our study extent, we can impute any missing trait values uUsing the phylogeny from Jetz & Pyron 2019. Based on the missing trait data of our focal species, we should either use Rphylopars or BHPMF + phylogeny to impute the missing trait data. I am going to use all the trait data available across the entire Southeast US anuran species group (n = 37 species) to impute the data using Rphylopars.

```{r phylogeny, echo = F, fig.height=4.5}
taxa_table<-trait_used %>% dplyr::select(species) %>%
  mutate(Tree.name.syn=case_when(species == 'Dryophytes squirellus'~ 'Hyla squirella',
                             species == 'Dryophytes gratiosus'~ 'Hyla gratiosa',
                             species == 'Dryophytes cinereus'~'Hyla cinerea',
                             grepl('Dryophytes',species)~gsub('Dryophytes','Hyla', species),
                             species == 'Lithobates areolatus'~ 'Rana areolata',
                             species == 'Lithobates catesbeianus'~'Rana catesbeiana',
                             species == 'Lithobates sevosus'~'Rana sevosa',
                             species == 'Lithobates sylvaticus'~ 'Rana sylvatica',
                             species == 'Lithobates sphenocephalus'~'Rana sphenocephala',
                             grepl('Lithobates',species)~gsub('Lithobates','Rana', species),
                             T~species),
         Tree.name=sub(' ','_',Tree.name.syn)) %>%
  filter(species != 'Lithobates kauffeldi')
# import phylogeny from Jetz and Pyron 2018
tree_orig <- read.tree(file = "input_data/amph_shl_new_Consensus_7238.tre") 
tree.pruned <- keep.tip(phy = tree_orig, 
                        tip = taxa_table$Tree.name) 
taxa_table$node<-nodeid(tree.pruned, taxa_table$Tree.name)
#plot(tree.pruned, font=2)
```

Using the trait_used dataframe (which selected regional traits over interspecific summarized values if available), we build a trait dataframe to do our inputation on.

```{r build trait table for phylo}
trait.df<-trait_used %>%
  select(-Tforage_optim) %>%
  mutate(category=case_when(species %in% mnm_spp & species %in% com_spp ~ 'both papers',
                            species %in% mnm_spp ~ 'mnm paper',
                            species %in% com_spp ~ 'comparison paper',
                            T ~ 'rare in SE')) %>%
  right_join(taxa_table, by='species') %>%
            mutate(tree.lab=paste0(substr(Tree.name,1,1), '. ',
                                   sub('.*\\_','',Tree.name))) %>%
  replace_na(list(n=0, comp=0))
#head(trait.df)
```
```{r plot the phylo tree, echo=F, warning=F, message=F}
trait_tree<-Rphylopars::left_join(tree.pruned, trait.df, by='node')
ggtree(trait_tree, aes(color=category, label=tree.lab), size=3)+ 
  geom_tiplab(as_ylab=T)
```


## Use Rphylopars to imput continuous trait data

We first make sure we are only including continuous traits in resulting dataframe and then get it in the format requested by phylopars. 

```{r trait mat, warning=F, message=F}
names(trait.df[,c(16, 6:13)])
trait.mat<-trait.df[, c(16, 6:13)] %>%
  rename('species'='Tree.name')
```

We use the following code to impute the traits using the *Rphylopars* package. 
- estimates phylogenetic and phenotypic variance-covariance matrices 
- assumes phenotypic covariance is equivalent among species
- uses a Browning motion model of evolution
Essentially is a maximum likelihood frequentist way to impute traits

```{r phylogenetic models, echo=T}
p_BM <- phylopars(trait_data = data.frame(trait.mat), tree = tree.pruned,
                  phylo_correlated=F, model='BM')
p_lamda <- phylopars(trait_data = data.frame(trait.mat), tree=tree.pruned,
                     phylo_correlated=F, model= "lambda")
p_OU <- phylopars(trait_data = data.frame(trait.mat), 
                  tree = force.ultrametric(tree.pruned),
                  phylo_correlated=F, model='OU')
```


```{r phyomodel out, echo=F}
kable(data.frame(model=c('lambda','Brownian motion','OU'),
                 df=c(p_lamda$npars, p_BM$npars, p_OU$npars),
                 loglik=round(c(p_lamda$logLik, p_BM$logLik, p_OU$logLik),2)) %>%
        mutate(AIC =round(c(AIC(p_lamda),AIC(p_BM), AIC(p_OU)), 2)) %>%
        arrange(AIC),
      caption='Best model for phylogenetic imputation')
```

```{r}
summary(p_OU)
```

Reviewer requested some test of imputation consistency. 
```{r phyloreg loo}
# Code from https://github.com/ericgoolsby/Rphylopars/issues/51

# imputed_ind = individual-level imputations
# imputed_sp = species-level imputations
imputed_sp <- trait_data <- data.frame(trait.mat)
imputed_sp[,-1] <- NA
for(i in 1:nrow(trait_data))
{
  for(j in c(1:4, 8))
  {
    temp_trait_data <- trait_data
    temp_trait_data[i,j+1] <- NA
    p_temp <- phylopars(trait_data = temp_trait_data,
                        tree = force.ultrametric(tree.pruned),
                  model='OU', phylocov_fixed = p_OU$pars$phylocov)

    imputed_sp[i,j+1] <- p_temp$anc_recon[trait_data$species[i],j]
  }
}

sp_mean_abs_error <- colMeans(imputed_sp[,-1] - trait_data[,-1],na.rm=TRUE)
sp_mean_sq_error <- colMeans((imputed_sp[,-1] - trait_data[,-1])^2,na.rm=TRUE)

data.frame(traits=colnames(trait_data)[-1],
          "sp bias"=round(sp_mean_abs_error,3),
          "sp error" = round(sp_mean_sq_error,3),
          n = colSums(!is.na(trait_data[,-1])), 
          minimum=sapply(trait_data[,-1], min, na.rm=T),
          maximum=sapply(trait_data[,-1], max, na.rm=T))

imputed_sp[,2] - trait_data[,2]
```


```{r results graph,  fig.height=4, echo=F}
#results!
imp_results <- p_OU$anc_recon[1:length(tree.pruned$tip.label),] %>%
  as_tibble() %>%
  bind_cols(species=rownames(p_OU$anc_recon[1:length(tree.pruned$tip.label),])) %>%
  left_join(trait.mat, by='species')%>%
  pivot_longer(-species) %>%
  mutate(variable=gsub('.y','', gsub('\\.x','', name)),
         impute=ifelse(grepl('y',name), 'No','Yes'))  %>%
  dplyr::select(-name) %>%
  pivot_wider(names_from = impute, values_from=value)  %>%
  mutate(Imputed=ifelse(is.na(No), 'Imputed', 'Literature'),
         genus=gsub("\\_.*", "\\1", species),
         family=case_when(genus %in% c('Acris','Hyla','Pseudacris') ~ 'Hylidae',
                          genus %in% c('Anaxyrus','Incilius') ~ 'Bufonidae',
                          genus %in% c('Lithobates','Rana') ~'Ranidae',
                          genus == 'Gastrophryne'~'Microhylidae',
                          genus == 'Scaphiopus' ~ 'Scaphiopodidae')) 
imp_results %>%
  ggplot(aes(x=family, y=Yes, color=Imputed))+
  geom_boxplot(color='black', fill=NA, outlier.alpha = 0)+
  geom_point(position=position_dodge(width=0.5), alpha=.5)+
  facet_wrap(~variable, scales='free_y')+
  labs(x='Family',y='Value')+
  scale_color_manual(values=c('forestgreen','goldenrod'))+
  theme_classic() +
  theme(legend.position='top',
        axis.text.x = element_text(angle=20, hjust=.87, vjust=1))
```


```{r inputation results table, echo=F}
kable(imp_results %>% 
  left_join(taxa_table, by=c('species'='Tree.name')) %>%
  filter(species.y %in% focal_spp) %>% count(species, Imputed) %>%
  pivot_wider(names_from = Imputed, values_from = n) %>%
  mutate(per=Imputed / c(Imputed + Literature)*100),
  caption='The number of MNM traits either imputed or from the literature, and the percentabe of imputed traits for each species')
```

```{r species discussion, echo=F}
kable(imp_results  %>%
        left_join(taxa_table, by=c('species'='Tree.name')) %>%
        rename(Tree.name=species, species=species.y) %>%
        filter(species %in% focal_spp) %>%
  group_by(species) %>%
  count(Imputed) %>% 
  filter(Imputed == "Imputed") %>%
  mutate(perimp=round(n/8,3)) %>%
  ungroup() %>% summarize(summary(n), summary(perimp)) %>%
    mutate(row.names=c('min','.25','median','mean','.75','max')) %>%
    select(row.names, everything()),
  col.names=c('parameter', 'N traits','% traits'),
  caption='Summary of imputed traits')
```

This gives us `r imp_results  %>% left_join(taxa_table, by=c('species'='Tree.name')) %>% rename(Tree.name=species, species=species.y) %>% filter(species %in% focal_spp) %>% group_by(species) %>% count(Imputed) %>% filter(Imputed == "Imputed") %>% ungroup() %>% summarize(sum(n)) %>% as.numeric()` traits (of a possible `r length(focal_spp)` x 8 = `r length(focal_spp)*8`). 

```{r rcs focal sp plot, echo=F, fig.height=2.75, fig.width=5}
rcs<-read_csv('C:/Users/Owner/Desktop/Git/National-RCS/Anuran_RCS_and_components_20220908.csv',
              show_col_types=FALSE) %>% 
  filter(grepl('North', extent)) %>%
  dplyr::select(species, genus, family, AOO_WS_sqkm, RCS_WS) %>%
  # manually adding in 'Lithobates sevosus' information
  bind_rows(data.frame(species='Lithobates sevosus', 
                       genus='Lithobates',family='Ranidae',
                       AOO_WS_sqkm=737.8, RCS_WS=1.0001)) %>%
  mutate(rankRCS=91-rank(RCS_WS))
imp_results %>%
  left_join(taxa_table, by=c('species'='Tree.name'))%>%
  dplyr::select(-species) %>% rename(species = species.y) %>%
  group_by(species) %>%
  summarize(n_imp=sum(Imputed=='Imputed'), .groups='drop') %>%
  left_join(rcs, by='species') %>%
  ggplot(aes(x=RCS_WS, y=n_imp))+
  geom_point()+
  geom_text_repel(data=. %>% filter(n_imp<5), aes(label=species), 
                  min.segment.length = .01, size=2.5)+
  labs(y='Number of Imputed Traits',
       x='RCS index - North American extent, watershed grain size')+
  theme_classic()
```

## Outputting the trait tables to query for MNMs

Below, I build two related trait datasets that incorporate traits from this triat imputation analysis, activity data from ATraiU-TaDS, and habitat data from ATraiU. One dataframe is for the MNM manuscript that contains the 13 species that meet our criteria for inclusion in the MNM manuscript (at least 3 of 4 important MNM traits). The second is MNM trait data for all 40 species considered within the SE US for the trait database compilation. The second will aid in expanding the comparison paper species list. 

```{r}
atraiu<-read_csv('C:/Users/Owner/Downloads/ATraiU/ATraiU_summary_values_2020AUG.csv', 
                 show_col_types = F)
used_traits1<-imp_results %>% #imputed thermoregulation traits
  select(species, variable, Yes, Imputed) %>%
  group_by(species) %>%
  # tally imputed traits to keep track
  mutate(imp.trait=paste(ifelse(Imputed=='Imputed', variable, ''), collapse=''),
         n.imp=sum(Imputed=='Imputed')) %>%
  select(-Imputed) %>%
  pivot_wider(names_from=variable, values_from=Yes) %>%
  # join in taxonomic information
  rename(Tree.name=species) %>% left_join(taxa_table, by='Tree.name') %>%
  # filter to only the focal species
  filter(species %in% c(focal_spp, "Lithobates sevosus")) %>% ungroup() %>%
  select(-node, -Tree.name.syn, -Tree.name) %>%
  # bring in ATraiU for burrowing & climbing
  left_join(atraiu %>% select(family, latin_name, s.fossorial, s.arboreal),
            by=c('species'='latin_name')) %>%
  # bring in categorical activity data
  left_join(read.csv('ATraiU 2.0/ATraiU2_summary_values_2022SEPT.csv') %>%
              dplyr::select(species, nocturnal, diurnal, crepuscular) %>%
              mutate(across(-species, ~ifelse(.x > 0,1,0))), by='species') %>%
  # reorder so column order is nicer
  select(family, species,imp.trait, n.imp, s.fossorial, s.arboreal, 
         nocturnal, diurnal, crepuscular, everything()) %>%
  # estimating activity traits for Lithobates areolatus
  mutate(nocturnal=case_when(species %in% c('Lithobates areolatus', "Pseudacris brachyphona",
                                            "Pseudacris fouquettei") ~ 1, T~nocturnal),
         diurnal=case_when(species %in% c("Pseudacris brachyphona", "Pseudacris fouquettei") ~ 0,
                           species == 'Lithobates areolatus' ~ 1,
                           T~diurnal),
         crepuscular = case_when(species %in% c("Pseudacris brachyphona", "Pseudacris fouquettei") ~ 0,
                                 species == 'Lithobates areolatus' ~ 1,
                                 T~crepuscular),
         s.fossorial = as.numeric(s.fossorial),
         s.arboreal = as.numeric(s.arboreal))
write.csv(used_traits1, 'input_data/inputed_physio_traits_011923.csv', row.names=F)

# writing out all imputed traits to explore options of expanding comparison species list 
p_OU$anc_recon[1:length(tree.pruned$tip.label),] %>%
   as_tibble() %>%
  bind_cols(species=rownames(p_OU$anc_recon[1:length(tree.pruned$tip.label),])) %>%
  mutate(species=gsub('_',' ', species)) %>%
  # bring in ATraiU for burrowing & climbing
  left_join(taxa_table %>% dplyr::select(species, Tree.name.syn),
            by=c('species'='Tree.name.syn')) %>%
  select(-species) %>% rename(species=species.y) %>%
  left_join(atraiu %>% select(latin_name, s.fossorial, s.arboreal),
            by=c('species'='latin_name')) %>%
  # bring in categorical activity data
  left_join(read.csv('ATraiU 2.0/ATraiU2_summary_values_2022SEPT.csv') %>%
              dplyr::select(species, nocturnal, diurnal, crepuscular) %>%
              mutate(across(-species, ~ifelse(.x > 0,1,0))), by='species') %>%
  # estimating activity traits for Lithobates areolatus
  mutate(nocturnal=case_when(species %in% c('Lithobates areolatus', "Pseudacris brachyphona",
                                            "Pseudacris fouquettei") ~ 1, T~nocturnal),
         diurnal=case_when(species %in% c('Lithobates areolatus', "Pseudacris brachyphona",
                                            "Pseudacris fouquettei") ~ 0, T~diurnal),
         crepuscular = case_when(species %in% c('Lithobates areolatus', "Pseudacris brachyphona",
                                            "Pseudacris fouquettei") ~ 0, T~crepuscular),
         s.fossorial = as.numeric(s.fossorial),
         s.arboreal = as.numeric(s.arboreal)) %>%
  # rearranging columns
  dplyr::select(species, everything()) %>%
  write.csv('input_data/MNM_traits_SEspp_011923.csv', row.names = F)
#which(is.na(used_traits1), arr.ind=T)
```

The figures below show the variation among taxa in major traits used in our mechanistic niche model. These interspecific traits, with variation in geographic location, will influence the outcomes of the MNMs. 

```{r, fig.width=5, fig.height=2, echo=F}
nictp<-theme_classic()+
  theme(axis.text.x = element_text(angle=25, hjust=.9, size=9))
plt.df<-used_traits %>% mutate(pltsp=paste0(substr(species, 1,1), '. ', 
                                            sub(".*? ", "", species)))
#mass
ggplot(plt.df, aes(x=family, y=Mass, label=pltsp)) +
  geom_point() + geom_text_repel(size=3)+
  scale_y_log10('Wet Mass, g', breaks=c(1, 10, 50, 100, 250))+nictp
```

```{r thermal traits, fig.width=7, fig.height=2.5, echo=F}
# thermal limits & scope
plt.df %>% select(family, pltsp, CTmax, CTmin) %>%
  mutate(CTscope=CTmax - CTmin) %>%
  pivot_longer(-c(family, pltsp)) %>%
  ggplot(aes(x=family, y=value, label=pltsp)) +
  geom_point() + geom_text_repel(size=2.5)+
  facet_wrap(~name, scales='free_y')+nictp
# tbask & tforage
plt.df %>% select(family, pltsp, Tbask, Tforage_min, Tforage_max) %>%
  pivot_longer(-c(family, pltsp)) %>%
  ggplot(aes(x=family, y=value, label=pltsp)) +
  geom_point() + geom_text_repel(size=2.5)+
  facet_wrap(~name, scales='free_y')+nictp
```
```{r other temp traits, fig.width=5, fig.height=2.5}  
# emergence temperature
plot_grid(ggplot(plt.df, aes(x=family, y=Tmerge, label=pltsp)) +
  geom_point() + geom_text_repel(size=3)+
    scale_y_continuous('Emergence Temperature, Celcius')+nictp,
# tpref
ggplot(plt.df, aes(x=family, y=Tpref, label=pltsp)) +
  geom_point() + geom_text_repel(size=3)+
  scale_y_continuous('Preferred Temperature, Celcius')+nictp)
rm(p_BM, p_lamda, p_OU, plt.df, 
   trait_filt, trait_restrict, trait.df, trait.mat,  trait_used, 
   usa_traits, usa,edge_sts, edge_sts2) #clean up space 
```


### Outputting Intraspecific Variance Trait Table

Now that the trait is decided and the species identified, I will output the subset species trait values to run though the function created in 4a_MNMAcrossSpaceCode.Rmd.

```{r}
trait_filt %>% filter(traitName == 'Tpref',
                      species %in% c('Acris blanchardi','Anaxyrus americanus',
                                     'Pseudacris crucifer', 'Lithobates clamitans')) %>% 
  select(species, traitName, traitValue, verbatimLocality, traitSource) %>% 
  arrange(species, verbatimLocality) %>% ungroup() %>%
  select(-traitName) %>% rename(Tpref=traitValue) %>%
  left_join(used_traits1 %>% select(-Tpref, -family, -imp.trait, -n.imp), by='species') %>%
  write.csv('input_data/intraspecific_variation_subset.csv', row.names = F)
```

```{r}
trait_filt %>%
  filter(traitName != 'Mass', !grepl('forage', traitName),
         species %in% mnm_spp) %>%
  group_by(traitName, species) %>%
  filter(traitValue == max(traitValue)|traitValue == min(traitValue)) %>%
  arrange(species, traitName)
used_traits1 %>%
  filter(species %in% c('Dryophytes chrysoscelis','Lithobates palustris','Lithobates sylvaticus'))
```

# Exploring accounting for Acclimation using data from ATraiU-TaDS

So one issue is that we know there is intraspecific variation in critical thermal maxima and other thermoregulation trait values. We could try to build curves to control for this acclimation-driven variation in CTmax. Essentially we would derive a function between acclimation temperature and CTmax for each species (if possible). We would then apply that function across a raster of mean temperature during the warmest quarter (WorldClim bioclim variable I think). Then, before running the ectotherm model, we would pull the acclimation-adjusted CTmax for that location. 

First, lets explore the information we have about acclimation ability in SE Anurans. We find that there are only six species that had different acclimation temperatures and CTmax values.

```{r}
all_traits<-read.csv('ATraiU 2.0/ATraiU2_full_2022SEPT.csv')
#head(all_traits)
CTmax_acc<-all_traits %>% 
  filter(traitName=='CTmax', 
         lifeStage %in% c('adult','unknown'),
         !is.na(traitAcclimation)) %>%
  mutate(traitAcclimation=case_when(traitAcclimation == '4C and 14C' ~ 14,
                                    traitAcclimation == '20 to 24' ~ 22,
                                    T ~ as.numeric(traitAcclimation)),
         traitValue = as.numeric(traitValue)) 
```

```{r spp with acc info, echo=F}
kable(CTmax_acc %>% 
  filter(species %in% focal_spp) %>%
  group_by(species) %>% 
  mutate(n=n(), range=max(traitAcclimation)-min(traitAcclimation)) %>%
  filter(n >1, range > 0) %>%
  summarize(meanCTmax=round(mean(traitValue),2), 
            sdCTmax=round(sd(traitValue),2), 
            accRange=mean(range), nPts=mean(n)),
  caption='Species with more than one acclimation temperature available for adult CTmax trait values',
  col.names = c('Species', 'Mean CTmax', 'sd CTmax', 'Acclimation temp. range', 'N CTmax values'))
```

That being said, it does look like there is some variation in acclimation ability. The lines below represent CTmax as a function of log(acclimation temperature) for quick analysis; the black line represents the function using all SE Anuran data. 

```{r acclimation plot, echo=F, fig.width=6, fig.height=2.75}
CTmax_acc %>% ggplot(aes(x=traitAcclimation, y=traitValue))+
  geom_smooth(method="lm", formula=y~log(x), se=50, color='black')+
  geom_smooth(data=. %>% filter(species %in% c('Anaxyrus fowleri',
                                               'Lithobates catesbeianus',
                                               'Lithobates clamitans',
                                               'Lithobates palustris',
                                               'Lithobates sylvaticus',
                                               'Scaphiopus holbrookii')),
              aes(color=species),
              method="lm", formula=y~log(x), se=0)+
  geom_text_repel(data=. %>% 
                    group_by(species) %>% 
                    filter(species %in% c('Anaxyrus fowleri', 'Lithobates catesbeianus',
                                          'Lithobates clamitans', 'Lithobates palustris',
                                          'Lithobates sylvaticus', 'Scaphiopus holbrookii'),
                                    traitAcclimation == max(traitAcclimation)) %>%
                    distinct(species, traitAcclimation, .keep_all=T),
                  aes(label=species), size=2.5,
                  min.segment.length = 0, nudge_x = 5)+
  geom_point(aes(color=species))+theme_classic()+
  scale_y_continuous('Measured Critical Thermal Maxima')+
  scale_x_continuous('Acclimation Temperature')+
  theme(legend.position='none')+
  ggtitle('Potential acclimation curves')
```

I think I don't have enough data to calculate species' level acclimation curve, even for species with more than two points. I apply the formula describbed in Comte & Olden 2017 to account for variation in CTmax when estimating warming tolerance across species ranges for both freshwater and marine fish. 
$$ CTmax = b + (a - b)e^x $$

where a is the intercept on the y axis, b is the asymptote (maximum CTmax), and x is c*Acclimation temperature (where c is a pseudo-acclimation rate). 

Below I calculate a single acclimation curve for all SE Anurans, acclimation curves where CTmax intercept, CTmax asymptote, and pseudo-acclimation rate vary among family individually, and an acclimation curve where both CTmax intercept and CTmax asymptote vary among families (Comte & Olden 2017 for fish). I could not get the models to converge when I used species as the random factor, even when restricting to the 6 species with multiple acclimation points (likely because median n = 4 points). 

```{r}
df<-CTmax_acc %>% dplyr::select(family, traitValue, traitAcclimation) %>%
  rename(y=traitValue, x=traitAcclimation)
#length(unique(df$family))

allSE<-nls(y~b+(a-b)*exp(c*x), data=df, start=c(a=10, b=40, c=-0.1))
asy<-nlme(y~b+(a-b)*exp(c*x), data=df, 
             random=b~1|family, fixed=a+b+c~1,
             start=c(a=10, b=40, c=-0.1),
             control=nlmeControl(pnlsTol=.1, msMaxIter=1e6, tolerance=1e-4))
int<-nlme(y~b+(a-b)*exp(c*x), data=df, 
             random=a~1|family, fixed=a+b+c~1,
             start=c(a=10, b=40, c=-0.1),
             control=nlmeControl(pnlsTol=10, msMaxIter=1e7, tolerance=1e-4))
rate<-nlme(y~b+(a-b)*exp(c*x), data=df, 
             random=c~1|family, fixed=a+b+c~1,
             start=c(a=10, b=40, c=-0.1),
             control=nlmeControl(pnlsTol=.1, msMaxIter=1e6, tolerance=1e-4))
asyint<-nlme(y~b+(a-b)*exp(c*x), data=df, 
             random=a+b~1|family, fixed=a+b+c~1,
             start=c(a=10, b=40, c=-0.1),
             control=nlmeControl(pnlsTol=1, msMaxIter=1e7, tolerance=1e-4))
```

We find that two related top models. CTmax asymptote definitely varies by family and the model that has CTmax asymptote + CTmax intercept vary among families is second best because of increased df penalization. The below plot uses the family-level asymptote and intercept models for the lines, with the CTmax's calculated by trait imputation marked by triangles and at acclimation temperature 31. Because these functions reduce interspecific variation within families, I am leaning towards not accounting for acclimation with this method for our analysis. 

```{r acc model outputs, echo=F}
kable(tibble(anova(asy, int, rate, asyint, allSE)) %>%
  dplyr::select(Model:logLik) %>%
  mutate(`Model name`=c('asy','int','rate','asyint','allSE')) %>%
  arrange(desc(logLik)),
  caption='Models relating acclimation temperature to CTmax')
summary(asy)
```

Using the above model, we could try to account for acclimation and reestimate CTmax at each point. But, CTmax doesn't seem to vary that much after an acclimation temperature of 11 degrees and it would potentially remove a lot of interspecific variation. 

```{r plot what acclimation looks like, echo=F, fig.width=6, fig.height=4}
sapply(predict(asy, 
        data.frame(x=rep(seq(5,30,1), each=5),
                   family=rep(unique(df$family), 26)), asList=T), '[') %>%
  as_tibble() %>% pivot_longer(everything()) %>%
  mutate(x=rep(seq(5,30,1), each=5)) %>%
  ggplot()+
  geom_point(data=CTmax_acc, aes(y=traitValue, x=traitAcclimation, color=family))+
  geom_line(aes(x=x, y=value, color=name))+
  geom_text_repel(data=used_traits %>%
              mutate(pltsp=paste0(substr(species, 1,1), '. ', 
                                            sub(".*? ", "", species))),
             aes(x=31, y=CTmax, label=pltsp),hjust=0, size=2,
             min.segment.length = 0)+
  geom_point(data=used_traits,
             aes(x=31, y=CTmax, color=family), alpha=0.8,  pch=2, size=2)+
  scale_y_continuous('CTmax temperature')+
  scale_x_continuous('Acclimation temperature', expand=c(0.15,.15))+
  theme_classic()
```

## Intraspecific variation in thermoregulation traits

For species where we do have intraspecific estimates of traits, we can compare the intraspecific trait variation with interspecific trait variation. 

```{r trait variation, echo=F}
kable(all_traits %>%
  filter(species %in% focal_spp,
         traitName == 'CTmax',
         lifeStage %in% c('unknown','adult')) %>%
  mutate(traitValue = as.numeric(traitValue)) %>%
  # only keep maximum CT for each population
  group_by(species, verbatimLocality) %>% 
  filter(traitValue == max(traitValue)) %>%
  # calculate min, median and max across populations
  group_by(species) %>%
  summarize(minCTmax=min(traitValue),
            medCTmax=median(traitValue),
            maxCTmax=max(traitValue)) %>%
  mutate(CTmax_range=maxCTmax-minCTmax) %>%
  arrange(desc(CTmax_range)) %>%
  filter(CTmax_range > 0))
```

We do this using an ANOVA. 

```{r}
intra.df<-all_traits %>%
  filter(species %in% focal_spp,
         !(traitName %in% c('Activity', 'Tforage_max','Tforage_min','Tbask', 'Mass')),
         lifeStage %in% c('unknown','adult')) %>%
  mutate(traitValue = as.numeric(traitValue)) %>%
  # only keep maximum CT for each population
  group_by(species, traitName, verbatimLocality) %>% 
  filter(traitValue == max(traitValue)) %>%
  group_by(species, traitName) %>%
  mutate(n=n()) %>% filter(n != 1) %>%
  ungroup()
intra.df %>% group_by(traitName) %>% summarize(n=n(), spp=n_distinct(species))
```

```{r}
anova(lm(traitValue~species, data=intra.df[intra.df$traitName=='CTmax',]))
anova(lm(traitValue~species, data=intra.df[intra.df$traitName=='CTmin',]))
anova(lm(traitValue~species, data=intra.df[intra.df$traitName=='Tmerge',]))
anova(lm(traitValue~species, data=intra.df[intra.df$traitName=='Tpref',]))
```
```{r intra variation comp, echo=F, fig.width=6.5, fig.height=3}
intra.df %>% 
  ggplot()+geom_boxplot(aes(y=species, x=traitValue))+
  facet_wrap(~traitName, scales='free', ncol=2)+
  scale_x_continuous(expression('Trait Value, temperature  ('*degree*C*')'))+
  theme_classic()+ theme(axis.text.y=element_text(face='italic'), strip.background=element_blank(),
                         strip.text = element_text(face='bold'))
```

I think it sufficient to say that, except for CTmin, intraspecific variance is less than interspecific variance and we are more confident in using interspecific traits for our MNMs.

# Downloading NCEP data 

The below code is by Mike Kearney and was reported in the NicheMapR google group as a way to increase the speed of microclimate model creation. I ran it for ~ 10 years and then uploaded the results onto the HPC.

```{r download ncep, eval=F}
# Code by Mike Kearney!
library(curl)
years <- seq(2010, 2011)
vars <- c('tmax.2m.gauss.','dswrf.sfc.gauss.', 'prate.sfc.gauss.', 'vwnd.10m.gauss.',
          'uwnd.10m.gauss.', 'air.2m.gauss.', 'tmin.2m.gauss.', 'shum.2m.gauss.',
          'pres.sfc.gauss.', 'dlwrf.sfc.gauss.', 'ulwrf.sfc.gauss.', 'tcdc.eatm.gauss.')
for(year in years){
  for(j in 1:length(vars)){
    curl_download(paste0("ftp://ftp.cdc.noaa.gov/Datasets/ncep.reanalysis2/gaussian_grid/",vars[j],year, ".nc"), 
                  paste0("../ncep_data/",vars[j],year, ".nc"),quiet=TRUE, mode="wb")   
  }
}
library(ncdf4)
dir2do <- "../ncep_data/"
setwd(dir2do)
files<-list.files()
files<-files[grep(".nc", x = files)]
vars4 <- c("tmin", "tmax", "air", "shum", "uwnd", "vwnd")
vars3 <- c("pres", "tcdc", "dswrf", "dlwrf", "ulwrf", "prate")
allvars <- c(vars4, vars3)
for(i in 1:length(files)){
  filenm <- files[i]
  orig <- nc_open(filenm)
  t<-ncvar_get(nc = orig, varid = "time", start = 1, count = -1)
  
  varid <- names(orig$var)[which(names(orig$var) %in% allvars)]
  
  if(varid %in% vars3){
    start <- c(1,1,1) # compose the start location vector for getting the data out of the ncdf file for this location
    count <- c(-1,-1,-1) # how much data are wanted? (-1 means all)
  }else{
    start <- c(1,1,1,1) # compose the start location vector for getting the data out of the ncdf file for this location
    count <- c(-1,-1,-1,-1) # how much data are wanted? (-1 means all)
  }
  
  data <- as.numeric(ncvar_get(orig, varid = varid, start = start, count)) # get the data
  lon<-ncvar_get(nc = orig, varid = "lon", start = 1, count = -1) # longitudes
  lat<-ncvar_get(nc = orig, varid = "lat", start = 1, count = -1) # latitudes
  t2 <- ncdim_def(name = "time", units = " hours since 1800-1-1 00:00:0.0", vals = t, unlim=FALSE) # define a time variable
  lon2 <- ncdim_def(name = "lon", units = "degrees_east", vals = lon, unlim=FALSE) # define a longitude variable
  lat2 <- ncdim_def(name = "lat", units = "degrees_north", vals = lat, unlim=FALSE) # define a latitude variable
  att<-ncatt_get(orig, varid, attname=NA, verbose=FALSE) # get attributes from original file
  if(varid %in% vars4){
    level<-ncvar_get(nc = orig, varid = "level", start = 1, count = -1) # levels
    level2 <- ncdim_def(name = "level", units = "m", vals = level, unlim=FALSE) # define a latitude variable
  }
  
  
  if(varid %in% vars3){
    var1 <- ncvar_def(name = varid, units = att$units, dim = list(lon2,lat2,t2), missval = att$missing_value, longname = att$long_name, prec='float') # define the variable
  }else{
    var1 <- ncvar_def(name = varid, units = att$units, dim = list(lon2,lat2,level2,t2), missval = att$missing_value, longname = att$long_name, prec='float') # define the variable
  }
  
  vars <- list(var1) # get variable definition
  file2<-gsub('.nc', replacement = '_time2.nc', x = filenm) # make a new file name
  file3<-gsub('.nc', replacement = '_time.nc', x = filenm) # make a new file name
  ncnew <- nc_create(file2,vars) # create a new ncdf file to put the data in
  ncvar_put(ncnew,var1,data) # put the data in
  nc_close(ncnew) # close the new file
  if(varid %in% vars3){
    cmd<-paste0("ncpdq -a lon,lat,time ",file2," ",file3) # command for ncpdq - permute to reverse order
  }else{
    cmd<-paste0("ncpdq -a lon,lat,level,time ",file2," ",file3) # command for ncpdq - permute to reverse order
  }
  
  system(cmd) # run ncpdq
  
  file.remove(file2) # delete the temporary file
}
```

I was concerned that longer time frames would make creating models more difficult. When the model can converge, more years does not significantly change the time needed to create the model, only the size of the resulting object. 

```{r length microclim, fig.width=6, fig.height=1.5, echo=F}
microtest<-read.csv('micro_20221111.csv')
plot_grid(microtest %>% 
  mutate(obj.size.n=as.numeric(gsub(' Mb', '', obj.size))) %>%
  ggplot()+
  geom_col(aes(x=nyear, y=obj.size.n))+
    scale_y_continuous('Size of model\nMb')+
  scale_x_continuous('Model temporal length, years', breaks=1:9)+
    theme_bw(),
  microtest %>% 
  ggplot()+
  geom_col(aes(x=nyear, y=totaltime))+
    scale_y_continuous('Time to build model\nminutes')+
  scale_x_continuous('Model temporal length, years', breaks=1:9)+
    theme_bw())
```

Downloading terra data for the climate extrapolations can be found here: https://groups.google.com/g/nichemapr/c/DC8dI5-0GXc/m/zF8fL357BQAJ

# Building points at which to run the MNMs

Now that I have the species pool, I identify points to run the models across a species range. The code chunk below creates a grid across the southeastern US that we will use to get the points we will use as inputs into the microclimate model. I will populate a grid at the resolution of our microclimate model on the high power computer (not enough memory on my laptop). First, I need to build range estimates to populate this grid. I do this by using the point occurrences from the RCS analysis, removing points outside 100km of the species' IUCN range (when available), adding a 5km buffer, and taking the union of all points to get an area of occupancy for each frog species. 

```{r build grid, eval=F}
library(tidyverse); library(sf); library(maps);library(stars)
# Build Points at which to run MNMs ------------
# study extent
se <- readRDS('b_RegionalRCS/spatial_extent.rds') %>%
  st_as_sf()
st_crs(se)

# create raster of study extent
se_r<-st_rasterize(se, dx=1, dy=1) %>% #turn into a raster of 1x1 kilometers
  st_as_sf(as_points = FALSE, merge = FALSE) %>%  #turn into a polygon
  rowid_to_column() # add an id column
st_area(se_r[1,])
cat('raster created\n')

se_grid<-NULL
focal_spp<-read.csv('AnuranMNMs/inputed_physio_traits_011923.csv') %>% pull(species)
focal_spp
for(s in focal_spp[focal_spp != 'Lithobates sevosus']){
  spp_aoo<-read_sf('/home/tracidubose/RCS_resub/L48_1km', layer=s) %>%
    st_buffer(4)
  spp_grid<-st_join(se_r, spp_aoo) %>% filter(!is.na(FID)) %>%
    mutate(species=s)
  se_grid <- bind_rows(se_grid, spp_grid)
  print(s)
}

# adding some points for L. sevosus
ls_aoo <- read_sf('RCS_Anuran_input/ANURA') %>%
  filter(binomial=='Lithobates sevosus',
         grepl('Extant', legend)) %>%
  st_transform(st_crs(spp_aoo))
head(ls_aoo)
se_grid %>% as_tibble() %>% count(species) %>% arrange(n) %>% slice(1)
# because the n grid should still be less thant the smallest aoo in other species, 
# not adding any buffer
spp_grid <- st_join(se_r, ls_aoo %>% select(binomial, yrcompiled)) %>% filter(!is.na(yrcompiled)) %>%
  mutate(species = 'Lithobates sevosus') %>%
  select(-yrcompiled, -binomial)
se_grid <-bind_rows(se_grid, spp_grid)
st_area(se_grid[1,])
se_grid %>% as_tibble() %>% count(species) %>%
  write_csv('AnuranMNMs/grid_count_species.csv')

# identify all species in a single grid
se_grid_spp <- se_grid %>% as_tibble() %>% # avoiding summarize on sf
  group_by(rowid) %>%
  summarize(spp_all=paste(unique(species), collapse=', '),
            n_spp=n())
nrow(se_grid_spp)
mnm_spp <- c('Lithobates catesbeianus', 'Lithobates clamitans', 'Anaxyrus americanus', 
             'Lithobates sylvaticus', 'Lithobates palustris', 
    'Pseudacris crucifer', 'Dryophytes versicolor', 'Acris blanchardi', 'Dryophytes chrysoscelis', 
    'Anaxyrus fowleri', 'Scaphiopus holbrookii', 'Gastrophryne carolinensis', 'Lithobates areolatus')
mnm_grid_spp <- se_grid %>% as_tibble() %>% # avoiding summarize on sf
  filter(species %in% mnm_spp) %>%
  group_by(rowid) %>%
  summarize(spp_all=paste(unique(species), collapse=', '),
            n_spp=n())

com_spp <- c("Acris gryllus", "Acris crepitans",
    "Anaxyrus terrestris", "Anaxyrus fowleri", "Anaxyrus quercicus", 
    "Dryophytes avivoca",  "Dryophytes andersonii", "Dryophytes cinereus", 
    "Dryophytes chrysoscelis", "Dryophytes femoralis", "Dryophytes gratiosus",
    "Dryophytes squirellus",
    "Gastrophryne carolinensis",
    "Lithobates capito", "Lithobates grylio", "Lithobates okaloosae", 
    "Lithobates heckscheri", "Lithobates sphenocephalus", "Lithobates areolatus", 
    "Lithobates virgatipes", "Lithobates sevosus"
    "Pseudacris ornata", "Pseudacris feriarum", "Pseudacris brachyphona",
    "Pseudacris fouquettei",
    "Scaphiopus holbrookii")
com_grid_sp <- se_grid %>% as_tibble() %>% # avoiding summarize on sf
  filter(species %in% com_spp) %>%
  group_by(rowid) %>%
  summarize(spp_all=paste(unique(species), collapse=', '),
            n_spp=n())
print(paste(com_grid_sp %>% filter(grepl('sevos', spp_all)), '\n'))
library(cowplot)
plot_grid(mnm_grid_spp %>% ggplot()+geom_histogram(aes(x=n_spp), binwidth = 1)+
            geom_text(stat="bin", aes(x=n_spp, label=..count.., y=..count..+5100), binwidth=1)+
            theme_classic()+scale_y_continuous('Count Grid Cells', labels = scales::comma)+
            scale_x_continuous('n Species', expand = c(0,0))+
            ggtitle('MNM species list'),
          com_grid_sp %>% ggplot()+geom_histogram(aes(x=n_spp), binwidth = 1)+
            geom_text(stat="bin", aes(x=n_spp, label=..count.., y=..count..+5100), binwidth=1)+
            theme_classic()+scale_y_continuous('Count Grid Cells', labels = scales::comma)+
            scale_x_continuous('n Species', expand = c(0,0))+
            ggtitle('CCVA Comparison species list'),
          ncol=1)
ggsave('AnuranMNMs/count_of_gridcells.jpg', width=6.5, height=4)


# get the centroid of each grid cell as a point
pts<-se_r %>% st_centroid() %>%
  left_join(se_grid_spp, by='rowid') %>%
  filter(!is.na(n_spp))

se_r %>%
  left_join(se_grid_spp, by='rowid') %>%
  filter(!is.na(n_spp)) %>%
  ggplot()+
  geom_sf(data=se)+
  geom_sf(aes(fill=n_spp, color=n_spp))+
  scale_color_viridis_c('N\nspecies', aesthetics=c('fill','color'))+
  theme_bw()
ggsave('AnuranMNMs/n_total_spp_raster.jpg', width = 5, height = 5)
# convert the grid of points to a dataframe for NicheMapR
# locational input into the microclimate function below
se_pts_df <- pts %>%
  st_transform(st_crs("+proj=longlat +ellps=WGS84 +datum=WGS84")) %>%
  st_coordinates() %>%
  bind_cols(rowid = pts$rowid, spp_all=pts$spp_all) 

se_pts_df %>% left_join(mnm_grid_spp %>% rename(mnm_spp=spp_all, mnm_n_spp=n_spp), by='rowid') %>%
  left_join(com_grid_sp %>% rename(com_spp=spp_all, com_n_spp=n_spp), by='rowid') %>%
write.csv(paste0('AnuranMNMs/points_ran_', format(Sys.Date()), '.csv'), row.names = F)
```

To investigate the amount of a species' latitudinal gradient that is represented in our region, we rasterize species IUCN range polygons and our 5km buffered estimates of species occurrence at a .5 decimal degree spatial resolution and plot the resulting latitudinal distribution. This figure is found within the supplement of our paper. 

```{r, fig.height=2.7, fig.width=6.5}
# bring in the IUCN range shapefile to filter point occurrences
iucn_all <- read_sf('../ANURA') %>% 
  filter(binomial %in% focal_spp, # only keep species we are interested in
         !grepl('Extinct',legend)) %>% # only keep ranges that are extant or introduced
  group_by(binomial) %>% summarize()
latlist_iucn<-NULL
for(i in iucn_all$binomial){
  latlist_iucn<-bind_rows(latlist_iucn,
            iucn_all %>%
    filter(binomial == i) %>%
    st_rasterize(dx=.5, dy=.5) %>%
   # st_as_sf(as_points = TRUE, merge = FALSE) %>%
    st_coordinates() %>%
    as_tibble() %>% mutate(species=i))
} 

na_r<-st_bbox(c('xmin'=-140, 'ymin'=0, 'xmax'=-40, 'ymax'=70))%>%
  st_as_sfc() %>% st_sf(crs=st_crs("+proj=longlat +ellps=WGS84 +datum=WGS84")) %>%
  st_rasterize(dx=.5, dy=.5) %>%
  st_as_sf(as_points = FALSE, merge = FALSE) %>%  #turn into a polygon
                       rowid_to_column()
iucnrng<-st_join(na_r, iucn_all) %>% filter(!is.na(binomial)) %>% st_centroid()
latlist_iucn <-st_coordinates(iucnrng) %>% 
  bind_cols(species=iucnrng$binomial)
rm(iucnrng)
se_dl <- se %>% 
  st_transform(st_crs("+proj=longlat +ellps=WGS84 +datum=WGS84")) %>%
  st_rasterize(dx=.5, dy=.5) %>% #turn into a raster of 1x1 kilometers
  st_as_sf(as_points = FALSE, merge = FALSE) %>%  #turn into a polygon
                       rowid_to_column()

# then we pull in each species' occurrence points from GBIF & HerpMapper
anuran_pts<-NULL
for(s in focal_spp[!grepl('sevos', focal_spp)]){
  anuran_pts<-bind_rows(anuran_pts,
                        read.csv(paste0('../National-RCS/data/occ_data/',s,'_20210330.csv')) %>% 
                        # only keep this information about the occurrence point
                          dplyr::select(source, species, Longitude, Latitude,
                                        coordinateUncertaintyInMeters, coordinatePrecision, elevation, 
                                        day, month, year))
}
pt1<-anuran_pts %>%  
  st_as_sf(coords=c('Longitude','Latitude'), 
           crs=st_crs("+proj=longlat +ellps=WGS84 +datum=WGS84")) %>%
  st_join(se_dl) %>% filter(!is.na(rowid)) %>%
  distinct(species, rowid, .keep_all = T)
latlist_occs<-st_coordinates(pt1) %>% bind_cols(species=pt1$species)
```

```{r range lat density graph, echo=F, fig.height=3, fig.width=6.5}
max.lat<-st_bbox(se %>% st_transform(st_crs("+proj=longlat +ellps=WGS84 +datum=WGS84")))[4]
min.lat<-st_bbox(se %>% st_transform(st_crs("+proj=longlat +ellps=WGS84 +datum=WGS84")))[2]
latlist_iucn %>%
            mutate(dist='IUCN range') %>%
            bind_rows(latlist_occs %>% 
                        mutate(dist='occurrences')) %>%
  filter(species %in% mnm_spp) %>%
  left_join(rcs %>% distinct(family, species), by='species') %>%
   mutate(pltsp=paste0(substr(species, 1,1), '. ', sub(".*? ", "", species))) %>%
  ggplot()+
  geom_density(aes(x=Y, fill=dist, color=dist), alpha=0.8)+
  geom_vline(xintercept=c(min.lat,max.lat))+
  scale_x_continuous('Latitude')+
  scale_y_continuous('Proportion of Range', breaks=c(0,.1,.2,.3,.4))+
  #scale_color_manual('species', values=c(met.brewer('Renoir'), 'grey'),
  #                   aesthetics = c('fill','color')) +
  #facet_grid(dist~family)+
  facet_wrap(~species) +
  scale_fill_grey('Source of Range', labels=c('IUCN range','5km buffered point occurrences'),
                  aesthetics=c('color','fill'))+
  theme_classic()+theme(axis.title=element_text(size=9),
                        axis.text=element_text(size=9),
                        legend.position = c(.8,.1),
                        legend.background =  element_rect(color='black',fill=NA),
                        legend.title=element_text(size=8),
                        legend.text = element_text(size=8),
                        strip.text=element_text(face='italic'),
                        strip.background = element_blank())
  #guides(fill=guide_legend(nrow=3, size=.5, alpha=1))
ggsave('results/FigS2_2_propdist_lat.jpg', height=4.5, width=6.5)
latlist_iucn %>%
            mutate(dist='IUCN range') %>%
            bind_rows(latlist_occs %>% 
                        mutate(dist='occurrences')) %>%
  filter(species %in% com_spp)%>%
  #left_join(rcs %>% distinct(family, species), by='species') %>%
   mutate(pltsp=paste0(substr(species, 1,1), '. ', sub(".*? ", "", species))) %>%
  ggplot()+
  geom_density(aes(x=Y, fill=dist, color=dist), alpha=0.8)+
  geom_vline(xintercept=c(min.lat,max.lat))+
  scale_x_continuous('Latitude')+
  scale_y_continuous('Range Probability Density', labels=function(x)round(x, digits=2))+
  #scale_color_manual('species', values=c(met.brewer('Renoir'), 'grey'),
  #                   aesthetics = c('fill','color')) +
  #facet_grid(dist~family)+
  facet_wrap(~pltsp, ncol=4, scales = 'free_y') +
  scale_fill_grey('Source of Range', labels=c('IUCN range','5km buffered point occurrences'),
                  aesthetics=c('color','fill'))+
  theme_classic()+theme(axis.title=element_text(size=9),
                        axis.text=element_text(size=9),
                        legend.position = c(.8,.01),
                        legend.background=element_rect(color='black', fill=NA),
                        legend.title=element_text(size=8),
                        legend.text = element_text(size=8),
                        strip.text=element_text(face='italic', size=9),
                        strip.background = element_blank())
  #guides(fill=guide_legend(nrow=3, size=.5, alpha=1))
ggsave('../CompareAnuranCCVA/results/FigS1_propdist_lat.jpg', height=6.5, width=6.5)
rm(iucn_all, latlist_occs, na_r)
```
```{r percent dist in SE, echo=F}
kable(latlist_iucn %>%
  group_by(species) %>% summarize(percent.dist.below.max.lat=round(ecdf(Y)(max.lat)*100,2)) %>%
  arrange(percent.dist.below.max.lat),
  caption='Percent of species IUCN range, gridded at a .5 decimal degree resolution, below the maximum latitude of our study region',
  col.names = c('species','% below 34.46 N'))
```
