==> 3_OnePtAnalysis.Rmd <==
---
title: "One-Point Interspecific Comparison"
author: "TP DuBose"
date: "11/18/2022"
output: pdf_document
---

```{r setup, include=FALSE}
library(tidyverse);library(sf);library(stars); library(knitr)
library(ggrepel); library(cowplot); library(ggspatial)
library(NicheMapR)
se<-st_as_sf(readRDS('../CompareAnuranCCVA/spatial_extent.rds')) %>%
  st_make_valid() %>%
  st_transform(st_crs("+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=37.5 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=km +no_defs"))
library(MetBrewer)
sppal<-read.csv('input_data/inputed_physio_traits_11152022.csv') %>%
  distinct(family, species) %>%
  mutate(col=case_when(family == 'Scaphiopodidae'~met.brewer('Lakota')[2],
                       family == 'Hylidae'~ met.brewer('Lakota')[1],
                       family == 'Bufonidae'~met.brewer('Lakota')[6],
                       family == 'Microhylidae'~met.brewer('Lakota')[3],
                       family == 'Ranidae'~met.brewer('Lakota')[4])) %>%
  group_by(family)%>%
  mutate(shp=c(1,2,3,6,7,8)[1:n()],
         pltsp=paste0(substr(species, 1,1), '. ', sub(".*? ", "", species))) %>%
  arrange(pltsp)
tmplab<-function(x){paste0(x, '°C')}
```

To investigate thermal sensitivity of anurans using mechanistic niche models (MNMs), we will need to run models at multiple points and then summarize the models at the specific levels of biological organization. 

Ideally this will be at the resolution of the environmental datasets. The NicheMapR team has explored how ordinary kringing can be used to interpolate microclimate data to a physiological parameter (article link here)[https://onlinelibrary-wiley-com.ezproxy.lib.vt.edu/doi/10.1111/jbi.13152]. In 2020, there was a discussion of using rasters to speed processing for large geographic extents (thread here)[https://groups.google.com/g/nichemapr/c/FNzXsigRdmk/m/7z_4-rBaBgAJ]

Below, we run a for loop to run a MNM at a point where each species' critical thermal maximum was determined. This gives us information about the interspecific difference in thermal sensitivity and controls for acclimation (because the acclimation at that point would be incorporated within the measurement of CTmax). This for loop is similar to the one that we then use to run across the entire Southeast United States to approximate intraspecific thermal sensitivity.

The for loop consists of two main parts: the microclimate model (micro_ncep) and then a for loop for different biophysical models (ectotherm). For the one point interspecific comparison, the microclimate model runs at a single point where the CTmax was determined (based on the verbatim locality listed in ATraiU-TaDS). If the microclimate model doesn't converge, it tries a higher integrator error tolerance. This error is usually thrown because of difficulties calculating snow presence and effencts on soil temperatures. Michael Kearny said that he seees pretty reliable results as long as the error tolerance isn't above 10 (google group answer)[https://groups.google.com/g/nichemapr/c/RQWCoW38w08/m/7c3b9xlFAAAJ]. The for loop then finds another for loop that implements different thermoregulatory parameter for each CTmax found at that point and saves summarized output for each point. The intraspecific for loop has a similar structure but is run on the high power computer (to allow for parallelization). First, within the raster we decide, it runs a microclimate model for that grid cell (with error tolerance code included). It then runs a for loop for a unique biophysical model that uses the thermoregulatory trait values for each species found within that grid cell. It saves the same output as the one point interspecific comparison for loop. The code for the intraspecific MNM model output generation can be found in [ file name here ].

# One Point Interspecific Comparison

I georeferenced verbatim localities for CTmax trait values from ATraiU-TaDS using Wikipedia/GeoHack (e.g., (Long Island, NY georeference)[https://geohack.toolforge.org/geohack.php?pagename=Long_Island&params=40.8_N_73.3_W_scale:10000000000000000000_source:dewiki]). Most of these points exist outside the Southeast US, thus why we need to run the models within the SE US even though accounting for acclimation will be difficult. 

```{r map of single points, warning=F, message=F, echo=F, fig.height=2.5, fig.width=4}
pt_locs<-read.csv('input_data/point_locations.csv') # georeferenced CTmax localities
traits<-read.csv('input_data/inputed_physio_traits_011923.csv') # traits used throughout our analyses
usa<-st_as_sf(map('usa','main', fill=T, plot=F))
can<-st_as_sf(map('world','canada', fill=T, plot=F)) %>% st_transform(st_crs(usa))
mex<-st_as_sf(map('world','mexico', fill=T, plot=F)) %>% st_transform(st_crs(usa))
pt_locs %>% 
  mutate(verbatimLocality=recode(verbatimLocality, 'Long Island, NY'='New York, Long Island')) %>%
  distinct(species, verbatimLocality, .keep_all = T) %>%
  st_as_sf(coords=c('longitude','latitude'), crs=st_crs("+proj=longlat +ellps=WGS84 +datum=WGS84")) %>%
  mutate(pltsp=paste0(substr(species, 1,1), '. ', sub(".*? ", "", species))) %>%
  group_by(verbatimLocality) %>%
  summarize(pltspall=paste(unique(pltsp), collapse='\n')) %>%
  ggplot() +
  geom_sf(data=mex, aes(geometry=geom)) +
  geom_sf(data=can, aes(geometry=geom)) +
  geom_sf(data=usa, aes(geometry=geom)) +
  geom_sf(data=se, fill='lightgrey')+
   geom_label_repel(aes(label = pltspall, geometry = geometry),
    stat = "sf_coordinates", size=1.9, fontface='italic',
    min.segment.length = 0, label.r=.1, label.size=.01, max.overlaps=20)+
  geom_sf(color='red', size=2)+
  theme_bw()+
  theme(axis.title=element_blank())+
  annotation_north_arrow()+
  coord_sf(ylim=c(25,52), xlim=c(-125, -65))
ggsave('results/FigS1_interspecific_trait_location.jpg', width=6.5, height=4)
```

```{r}
pt_locs %>% 
  mutate(verbatimLocality=recode(verbatimLocality, 'Long Island, NY'='New York, Long Island')) %>%
  distinct(species, verbatimLocality, .keep_all = T) %>%
  st_as_sf(coords=c('longitude','latitude'), crs=st_crs("+proj=longlat +ellps=WGS84 +datum=WGS84")) %>%
  st_filter(st_transform(st_buffer(se,10), st_crs(.)))
```


Below is the code for the one point interspecific comparison. It runs a microclimate model for each of the points above and for all the CTmaxes determined at that point. 

```{r hpc one point, eval=F}
# bash script to programatically run Mechanistic Niche models ----
# by Traci DuBose, last edited 11/22/2022

# set up the libraries
.libPaths(.libPaths()[3:1])
#library(devtools)
#devtools::install_github('mrke/NicheMapR')
#devtools::install_github('ilyamaclean/microclima')
library(tidyverse); library(maps);  library(sf); library(parallel); library(hoardr); library(stars); library(NicheMapR)

# DATA INPUTS ------------
coress<-5

# paths
PATH <- "./" #arc
PATH_out <- paste0(PATH, "OnePtResults/") # where to save the output
# traits used to parameterize the ectotherm function
traits <- read.csv(paste0(PATH, "inputed_traits.csv")) %>%
  mutate(across(c('s.fossorial', 's.arboreal', 'nocturnal','diurnal','crepuscular'),
                as.numeric))
focal_spp<-traits %>% pull(species) %>% unique() # focal species to run models for

# intraspecific CTmax traits
pt_loc<-read.csv(paste0(PATH, "point_locations_1pt.csv"))

# points at which to run the microclimate model
pts_df<-pt_loc %>%
  group_by(species, latitude, longitude) %>%
   tally() %>% arrange(desc(n))

# model parameters
start_date="01/01/2012"
end_date="31/12/2020"

cat('\ndata loaded\n')

# FUNCTION WHICH WE USE TO PARALLALIZE CODE ----------
# it should run the microclimate model and then an ectotherm model for each species found in that grid
anuranMNMs<-function(rw, write.micro.out=F){
  pt<-pts_df[rw,]
  cat(paste0(rw, ' row number \n', round(pt[1,3],3), '  ',round(pt[1,2], 3),'\n'))

  ecto_out <- NULL
  ERR <- 1.5
  micro <- micro_ncep(loc = c(as.numeric(pt[1,3])-0.00001,as.numeric(pt[1,2])+0.00001),
                     dstart = start_date, dfinish = end_date,
                     DEP = c(0, 2.5,  5,  10,  15,  20,  30,  50, 75,  100), #cm
                     minshade = 0, maxshade = 90, runshade=1,
                    # soilgrids=1,
                     spatial='./ncep_data/ncep_data',
                     dem.res=1000, # requested resolution DEM from elevatr in meters
                     Usrhyt = 0.01, # local height for organism
                     ERR = ERR,
                     run.gads=2)
  gc()
  while(min(micro$metout[,1])==0 & ERR <= 6){
        cat("model crashed, trying a higher error tolerance \n")
        ERR <- ERR + 0.5
        # rerun the microclimate with slightly higher error tolerance allowed
        micro <- micro_ncep(loc = c(as.numeric(pt[1,3])-0.00001,as.numeric(pt[1,2])+0.00001),
                     dstart = start_date, dfinish = end_date,
                     DEP = c(0, 2.5,  5,  10,  15,  20,  30,  50, 75,  100), #cm
                     minshade = 0, maxshade = 90, runshade=1,
                    #soilgrids=1,
                     spatial='./ncep_data/ncep_data',
                     dem.res=1000, # requested resolution DEM from elevatr in meters
                     Usrhyt = 0.01, # local height for organism
                     ERR = ERR,
                     run.gads=2)
        gc()
cat(paste('row',rw,'tried error ', ERR, 'and it worked ==', min(micro$metout[,1])!=0, '\n'))
      }
if(min(micro$metout[,1])==0){
mc.df<-data.frame(rowid=rw,
           lat=micro$longlat[2],
           long=micro$longlat[1],
           notes='broken model',
           err=ERR)
write.csv(mc.df, paste0(PATH_out, rw,'_microloc.csv'), row.names=F)
        }else{
mc.df<-data.frame(rowid=rw,
           lat=micro$longlat[2],
           long=micro$longlat[1],
           tRainfall=sum(micro$RAINFALL),
           aRainfall=sum(micro$RAINFALL)/(length(micro$dates2)/365),
           evel=micro$elev,
           slope=micro$SLOPE,
                      aspect=micro$ASPECT,
           err=ERR) %>%
mutate(across(-c(rowid, lat, long), ~round(.x, 3)))
write.csv(mc.df, paste0(PATH_out, rw,'_microloc.csv'), row.names=F)
bind_cols(ppt=micro$RAINFALL, dat=micro$dates2) %>% mutate(year=substr(dat,1,4)) %>%
        group_by(year) %>%
        summarize(mppt=mean(ppt), sdppt=sd(ppt), tppt=sum(ppt), .groups='drop') %>%
left_join(micro$metout %>%
             bind_cols(dat=micro$dates) %>%
             mutate(year=substr(dat, 1,4)) %>%
             group_by(year) %>% summarize(across(c(TALOC, TAREF), list(mean=mean, sd=sd, min=min, max=max)), .groups='drop'),
          by='year') %>%
mutate(across(-c(year), ~round(.x, 2)), rowid=rw) %>%
write.csv(paste0(PATH_out, rw, '_ppt_dat.csv'), row.names=F)
gc()}
  # ECTOTHERM MODEL STARTS -------
# get the rest of the trait data for the species at that point
spp_traits <- traits %>% filter(species == as.character(pt[1,1]))
cat(paste('ectotherm model for', pt[1,1], '\n'))

cat(paste(spp_traits,'\n'))
for(u in pt_loc[pt_loc$species == pt$species &
                   pt_loc$longitude == pt$longitude &
                   pt_loc$latitude == pt$latitude,"traitValue"]){
      ecto<-ectotherm(Ww_g=spp_traits["Mass"],
                      shape = 4, # shape based on leopard frog
                      CT_max=as.numeric(u),
                      CT_min=spp_traits["CTmin"],
                      T_pref=spp_traits["Tpref"],
                      # when is activity allowed
                      diurn=spp_traits["diurnal"],
                      nocturn=spp_traits["nocturnal"],
                      crepus=spp_traits["crepuscular"],
                      # can it go into a burrow or climb to cool off
                      burrow=spp_traits["s.fossorial"],
                      climb=spp_traits["s.arboreal"],
                      shdburrow = 1, #the animal's retreat is in the open (0), in the shade when above or below CTmin in sun (1) or in shade always
                      maxdepth = 10, #maximum depth of the burrow
                      T_F_min=max(c(as.numeric(spp_traits["Tforage_min"]), as.numeric(spp_traits["Tmerge"])), na.rm=T),
                      T_F_max=min(c(as.numeric(spp_traits["Tforage_max"]), as.numeric(spp_traits["CTmax"])), na.rm=T),
                      T_RB_min=spp_traits["Tmerge"],
                      T_B_min=spp_traits["Tmerge"],
                      # microclimate port from parent environment
                      nyears= micro$nyears,
                      minshades = micro$minshade,
                      maxshades = micro$maxshade,
                      alpha_sub = (1 - micro$REFL),
                      DEP = micro$DEP, KS = micro$KS, b = micro$BB,
                      PE = micro$PE, metout = micro$metout, shadmet = micro$shadmet,
                      soil = micro$soil, shadsoil = micro$shadsoil, soilmoist = micro$soilmoist,
                      shadmoist = micro$shadmoist, humid = micro$humid, shadhumid = micro$shadhumid,
                      soilpot = micro$soilpot, shadpot = micro$shadpot, tcond = micro$tcond,
                      shadtcond = micro$shadtcond, rainfall = micro$RAINFALL,
                      preshr = rep(101325 * ((1 -
                                                (0.0065 * as.numeric(micro$elev)/288))^(1/0.190284)),
                                   nrow(micro$metout)), elev = as.numeric(micro$elev),
                      longitude = as.numeric(micro$longlat[1]),
                      latitude = as.numeric(micro$longlat[2]))
bodytemps<-ecto$environ %>% as_tibble() %>%
    mutate(WT=u-TC) %>%
  group_by(YEAR, DOY) %>%
  summarize(activity0=sum(ACT == 0),
         activity1=sum(ACT == 1),
         activity2=sum(ACT == 2),
         shadeMean=round(mean(SHADE),2),
         across(c(WT,TC,TA,TSUB, TSKY, DEP), list(mean=mean, max=max, min=min)),
         nHoursAboveCTmax=sum(WT < 0),
         species=unique(pt$species),
         rowid=rw, .groups='drop') %>%
         mutate(across(c(ends_with('mean'), ends_with('min'), ends_with('max')), ~round(.x, 3))) %>%
  group_by(species, rowid, YEAR) %>%
  summarize(nDays=sum(nHoursAboveCTmax != 0),
        meannH=mean(nHoursAboveCTmax),
        WT_ptmin=min(WT_min),
        WT_min25=quantile(WT_min, .25),
        WT_25=quantile(WT_mean, .25),
        WT_75=quantile(WT_mean, .75),
        WT_ptmean=mean(WT_mean),
        nhrF=sum(activity2), .groups='drop',
        CTmax=u)
ecto_out<-bind_rows(ecto_out, bodytemps)
gc()
}
write.csv(ecto_out, paste0(PATH_out, rw, '_bodytemps.csv'))
return(ecto_out)
}

mclapply(1:nrow(pt_dfs), anuranMNMs, mc.cores=2)
```

Three files for each point are output: microclimate model basic information, annual climate information, and annual body temperature information. 

```{r}
PATH_onept<-'intermediate_results/one_pt/'
pts<-NULL
clim<-NULL
bt<-NULL
for(i in 1:24){
  pts <- bind_rows(pts, read.csv(paste0(PATH_onept, i, '_microloc.csv')))
  clim <- bind_rows(clim, read.csv(paste0(PATH_onept, i, '_ppt_dat.csv')))
  bt <- bind_rows(bt, read.csv(paste0(PATH_onept, i, '_bodytemps.csv')))
}

pts_df<-pt_locs %>%
  group_by(species, latitude, longitude) %>%
   tally() %>% arrange(desc(n)) %>%
  rowid_to_column()
```

```{r}
pts_df %>% 
  left_join(pts, by='rowid') %>%
  ungroup() %>% count(species)
nrow(pts)
```
 NOTE::: maybe figure out minimum value of soil grid???
```{r}
pt.bt.df<-left_join(bt %>% select(-X), clim %>% mutate(YEAR=year - 2012 + 1), by=c('rowid','YEAR'))
```

```{r}
pt.bt.df %>%
  ggplot()+
  geom_hline(yintercept=0, linetype='dashed', color='grey')+
  geom_line(data=. %>% group_by(species, year) %>% summarize(mWT=mean(WT_ptmin)),
            aes(x=year, y=mWT, group=species, color=species),
            position=position_dodge(width=.7))+
  #geom_point(aes(x=year, y=WT_ptmin, group=CTmax, color=species),
  #           position=position_dodge(width=.7))+
  stat_summary(aes(x=year, y=WT_ptmin, color=species), position=position_dodge(width=.7))+
  theme_classic()+
  theme(legend.position='bottom')

```

We can compare interspecific thermal sensitivity by identifying the species with the lowest median warming tolerance. The below graph does that. 

```{r }
pt.warmtol<-pt.bt.df %>%
  select(species, rowid, CTmax, year, WT_ptmin, WT_min25) %>%
   pivot_longer(cols=c('WT_min25','WT_ptmin')) %>%
  mutate(nm.F=factor(name, levels=c('WT_ptmin','WT_min25'),
                     labels=c('minimum','hottest quarter'))) %>%
  group_by(species, rowid, CTmax, nm.F) %>%
  summarize(medWT=median(value),
            WTmin2=quantile(value, .025),
            WTmin97=quantile(value, .975), .groups='drop') 
write.csv(pt.warmtol, 'intermediate_results/ptwarmtol_fig1adata.csv', row.names=F)
```

```{r}
pt.WTord<-pt.warmtol %>% group_by(species) %>% filter(nm.F=='minimum') %>%
  summarize(mWTmin=mean(medWT)) %>%
  select(species, mWTmin) %>% arrange(mWTmin) %>% pull(species)
pt.warmtol %>%
  mutate(spF=factor(species, levels=rev(pt.WTord))) %>%
  ggplot()+
  geom_hline(yintercept=0, linetype='dashed')+
  geom_linerange(aes(x=spF, ymin=WTmin2, ymax=WTmin97, color=CTmax, group=CTmax),
             position=position_dodge(width=.7))+
  geom_point(aes(x=spF, y=medWT, color=CTmax, group=CTmax), size=1.7,
             position=position_dodge(width=.7))+
  scale_y_continuous('Daily minimum warming tolerance', labels = tmplab)+
  scale_color_gradient(low='lightgrey',high='black', labels = tmplab)+
  theme_classic()+
  theme(axis.title.y=element_blank(),
        axis.text.y=element_text(face='italic'),
        legend.text = element_text(size=8),
        legend.title = element_text(size=10),
        legend.position='bottom',
        legend.margin=margin(t=-8))+
  coord_flip()+
  facet_grid(~nm.F, scales='free')
ggsave('results/FigS1_2_wt_results.jpg', width=5.5, height=6)
```


```{r}
sp.medWT<-pt.warmtol %>% select(species, rowid, CTmax, nm.F, medWT) %>% 
  pivot_wider(names_from=nm.F, values_from=medWT) %>%
  group_by(species) %>% summarize(across(c('minimum','hottest quarter'), median), .groups='drop') %>%
  left_join(read.csv('results/pt_wtmin_sum.csv') %>% filter(nm.F != 'median') %>%
              select(species, nm.F, medWT) %>% pivot_wider(names_from=nm.F, values_from=medWT) %>%
              rename(reg_min=minimum, reg25=`25th quartile`), by='species')
cor.test(sp.medWT$minimum, sp.medWT$reg_min, method='spearman')
cor.test(sp.medWT$`25th quartile`, sp.medWT$reg25, method='spearman')
sp.medWT %>% pivot_longer(-species) %>% mutate(type=factor(ifelse(grepl('hot',name)|grepl('25', name), 'hottest quarter', 'minimum'), levels=c('minimum','hottest quarter')),
                                              area=ifelse(grepl('reg', name), 'regional','acclimatization-control')) %>% select(-name) %>% 
  pivot_wider(names_from=area) %>%
  ggplot()+
  geom_point(aes(x=`acclimatization-control`, y=regional))+facet_wrap(~type, scales='free') + theme_classic()+
  theme(panel.background = element_rect(color='black', fill=NA),
        strip.background = element_blank(),
        axis.text=element_text(size=8))+
  scale_x_continuous(labels=tmplab, 
                     breaks=seq(0,15,by=5))+scale_y_continuous(labels=tmplab,breaks=seq(0,15, 5))
ggsave('results/FigS1_3_corWT.jpg', width=4, height=2)
```
