==> 4b_MNMAcrossSpaceResults.Rmd <==
---
title: "Results of the Mechanistic Niche Models Across Space"
author: "TP DuBose"
date: "11/22/2022"
output: pdf_document
---

```{r setup, include=FALSE}
library(tidyverse);library(sf);library(stars); library(maps); library(knitr)
#install.packages('ggridges')
library(ggridges)
library(ggrepel); library(cowplot); library(ggfortify)
library(emmeans)
se<-readRDS('../CompareAnuranCCVA/spatial_extent.rds')
library(MetBrewer)
tmplab<-function(x){paste0(x, '°C')}
latlab<-function(x){paste0(x, '° N')}
```


I ran the mechanistic niche models on the high power computer so that I could parallelize the code to run many models simultaneously and better sample across the Southeastern United States region. To run the points for the interspecific comparison, I call the *RunNicheMapR.bash* script, which calls two R scripts: one to run models (*pairedMCE_Rscript.r*) and one to collate results (*build_bodytemp_summary_df.r*). Because errors occur in the compilation of the microclimate model, I often have to run this script multiple times (> 16 times for anywhere from 1 hour to 8 hours each) to get through large chunks of points.

In the *pairedMCE_Rscript*, we first bring in three main data sets (traits, points to run, and points already completed), create a function over which to parallelize, and then a function to time out the FORTRAN error that occurs. The parallelized function takes one rowid (within the points to run file) as input, creates a microclimate model for that point, and then queries which species were found within 5km of that point. It then runs an ectotherm for each of those species and saves annual climate output, general climate output, and summarized body temperature for each anuran species in a single folder (results). 

While building the microclimate model, I run into an odd FORTRAN error and used the last answer on this [StackXchange question](https://stackoverflow.com/questions/7891073/time-out-an-r-command-via-something-like-try) to time out the process and prevent the core from hanging until the end of the SLURM job. It takes approximately 3 minutes to run a microclimate model ones and 20 seconds to run an ectotherm model. So I set the core time out to 10 minutes to be safe: 2 microclimate models and generous time for 15 species.

And then I use the *build_bodytemp_summary_df* script to collate and summarize all the completed models. It just uses a forloop to join all the files in the results folder into one file and outputs that file. It also outputs a stars object (i.e., pseudo raster) of all the species minimum daily minimum warming tolerance.

The following files are output from the above process. The plot plots the error tolerance used to build the models (MKearney says below 10 generally ok, above 6 not allowed).

```{r load files}
mnm_spp <- c('Lithobates catesbeianus', 'Lithobates clamitans', 'Anaxyrus americanus',
             'Lithobates sylvaticus', 'Pseudacris crucifer', 'Dryophytes versicolor', 
             'Acris blanchardi', 'Dryophytes chrysoscelis', 'Lithobates palustris', 
             'Anaxyrus fowleri', 'Scaphiopus holbrookii', 'Gastrophryne carolinensis',
             'Lithobates areolatus')
# all possible points
pts2run<-read.csv('intermediate_results/points_ran_2023-01-10.csv') %>%
  filter(!is.na(mnm_spp)) # removing rows that only have other spp in them
# body temperatures
ecto_out <- read.csv('intermediate_results/summarized_bodytemps_230127.csv') %>%
  filter(species %in% mnm_spp)
# location info
micro_out <- read.csv('intermediate_results/microclimate_info_230127.csv') %>%
  filter(rowid %in% unique(ecto_out$rowid)) # removing rows with only other species in them
# climate data at each point, not needed
#climout <- read.csv('intermediate_results/climate_data_230127.csv')

# make a dataframe that includes location and nice plotting name
ect.df<-ecto_out %>% 
  left_join(micro_out, by='rowid') %>%
   mutate(pltsp=paste0(substr(species, 1,1), '. ', sub(".*? ", "", species)))
# set up paths for intraspecific results too
PATH_intra_body <- 'intermediate_results/intra_summarized_bodytemps_230127.csv'
PATH_intra_loc <- 'intermediate_results/intra_microclimate_info_230127.csv'
```
```{r micro errors, echo=F, fig.width=4, fig.height=1}
micro_out %>%
  ggplot()+geom_histogram(aes(x=err), binwidth = .5)+
  scale_x_continuous('Microclimate Model Error Tolerance', breaks=1:4)+
  theme_classic()
```

# Intraspecific Trait availability

Below creates Figure 1, which plots the intraspecific traits available for CTmax, Tpref and Tmerge for our 13 species. It give the audience and idea of both intra and interspecific trait variation.

```{r trait scatter plot, echo=F, eval=F}
traits.df<-read.csv('input_data/intraspecific_variation_subset.csv')
read.csv('intermediate_results/regional_interspecific_traits.csv') %>% 
  filter(species %in% mnm_spp, lifeStage %in% c('adult','unknown'),
         traitName %in% c('CTmax','Tmerge','Tpref')) %>%
  select(species, traitName, traitValue) %>%
  mutate(type=ifelse(species %in% c('Acris blanchardi','Anaxyrus americanus',
                                    'Pseudacris crucifer', 'Lithobates clamitans') &
                       traitName == 'Tpref', 'intraspecific\ncomparison',
                     'available'), 
         traitValue=as.numeric(traitValue)) %>%
  bind_rows(read.csv('input_data/MNM_traits_SEspp_011923.csv') %>% filter(species %in% mnm_spp) %>%
              select(species, CTmax, Tpref, Tmerge) %>% 
              pivot_longer(-species, names_to='traitName', values_to = 'traitValue') %>%
    mutate(type='interspecific\ncomparison')) %>%
  mutate(typeF=factor(type, levels=c('available','intraspecific\ncomparison',
                                     'interspecific\ncomparison')),
         spF = factor(species, levels=rev(mnm_spp[order(mnm_spp)]))) %>% 
  distinct(species, traitValue, typeF, .keep_all=T) %>% #View()
  ggplot()+
  geom_point(aes(y=spF, x=traitValue, shape=typeF))+
  facet_wrap(~traitName, scales='free_x')+
  scale_x_continuous('Trait Value', labels=tmplab)+
  scale_shape_manual('Trait Use', values=c(3,4,16))+
  theme_classic()+
    theme(legend.title=element_text(size=9),
          legend.text=element_text(size=8),
          legend.margin = margin(l=-5, r=-5),
          axis.title.y=element_blank(),
          axis.title.x = element_text(size=9),
          axis.text.x=element_text(size=8),
          axis.text.y=element_text(size=8, face='italic'),
          panel.background = element_rect(color='lightgrey', fill=NA),
          strip.background = element_blank())
ggsave('results/Fig1_imp_trait_variance.jpg', width=6.5, height=3)
```

# Interspecific Comparisons

This gives us data for `r length(unique(ecto_out$rowid))` randomly sampled, but occupied points across our study region. This is `r round(length(unique(ecto_out$rowid))/nrow(pts2run), 4)*100`% of the `r nrow(pts2run)` occupied grid cells (1 km resolution).

```{r distribution of WT, echo=F, fig.width=6.6, fig.height=3}
ecto_out %>% dplyr::select(species, WT_25, WT_75, WT_ptmean) %>%
  filter(species != 'Lithobates areolatus') %>%
  pivot_longer(-species) %>%
  mutate(better.labs=case_when(name == 'WT_25' ~ '25th Quantile',
                               name == 'WT_75' ~ '75th Quantile',
                               name == 'WT_ptmean' ~ 'Mean')) %>%
  ggplot()+
  geom_density(aes(x=value, color=species)) +
  facet_wrap(~better.labs, scale='free_x') +
  scale_color_manual('species', values=c(met.brewer('Renoir'), 'lightgrey'))+
  theme_classic()+
  ggtitle('Daily Mean Warming Tolerance')+
  theme(title = element_text(size=10), 
        legend.position='bottom', legend.direction = 'vertical',
        legend.margin = margin(t=-10,l=-30),
        legend.text = element_text(size=8,  face='italic'),
        axis.text = element_text(size=9),
        axis.title = element_text(size=9))+
    guides(color=guide_legend(ncol=3, size=2))

ecto_out %>% dplyr::select(species, WT_ptmin) %>%
  ggplot()+
  geom_histogram(aes(x=WT_ptmin, fill=species), binwidth = .5) +
  geom_vline(xintercept=0)+
  scale_fill_manual('species', values=c(met.brewer('Renoir'), 'lightgrey'))+
  labs(x='Daily Mean Warming Tolerance', y='Count of Points')+
  theme_classic()+
  theme(legend.position='bottom', legend.direction = 'vertical',
        legend.margin = margin(t=-10,l=-30),
        legend.text = element_text(size=8,  face='italic'),
        axis.text = element_text(size=9),
        axis.title = element_text(size=9))+
    guides(fill=guide_legend(ncol=3, size=2))
```

Below I plot some summary figures: number of points run, number of points where minimum daily minimum WT is below 0, how many species that occurs for, plus the summary statistics for our three quantifications of daily minimum warming tolerance. 

```{r}
npts<-length(unique(ecto_out$rowid))
ecto_out %>% filter(WT_ptmin < 0) %>% 
  summarize(n=n_distinct(rowid), npts=n_distinct(rowid)/npts*100, nspp=n_distinct(species), 
            spp=paste(unique(species), collapse=', '))
summary(ecto_out$WT_ptmin)
summary(ecto_out$WT_min25)
summary(ecto_out$WT_min50)

```

```{r sample size mean table, echo=F}
ecto.summ<-ecto_out %>% group_by(species) %>%
  summarize(n_pts=n(), 
            minWT=median(WT_ptmin),
            WT25=median(WT_min25),
            WT50=median(WT_min50),
            mean_days=mean(nDays)) %>%
  mutate(across(-c(species, n_pts), ~round(.x, 2)))

kable(ecto.summ,
  caption='Sample size and summarized indicators of warming tolerance (WT) for each species',
  col.names = c('Species', 'N points', 'median daily minimum WT', 'median 25th daily min WT','median daily min WT', 'mean Days CTmax is exceeded'))
```

So we can first visualize each species result across space with the plot below. 

```{r species maps, echo=F, fig.width=6.5, fig.height=7, eval=F}
# need to run this on the super computer instead
ect.df %>%
  st_as_sf(coords=c('long','lat'),
           crs=st_crs("+proj=longlat +ellps=WGS84 +datum=WGS84")) %>%
  ggplot()+geom_sf(data=se %>% st_make_valid() %>% st_union())+
  geom_sf(aes(geometry=geometry, color=WT_ptmin, fill=WT_ptmin))+
  facet_wrap(~species)+
  scale_color_gradient2('Minimum\nWarming\nTolerance', 
                        #colors = met.brewer('OKeeffe1'),
                        aesthetics = c('fill','color'))+
  theme_void()+
  theme(strip.text = element_text(size=8))
ggsave('results/FigS2_tsm_results.jpg', width=6.5, height=6.5)
```

We already see some differences among species with G. carolinensis generally having large warming tolerance and L. palustris having a smaller warming tolerance.

```{r boxplots across space}
wtord<-ect.df %>% group_by(species) %>% 
  summarize(medWT=median(WT_ptmin)) %>% 
  arrange(medWT) %>% pull(species)

usa<-st_as_sf(map('usa','main',fill=T, plot=F)) 
se_inset<-ggplot()+
  geom_sf(data=usa, fill='white')+
  geom_sf(data=se, fill='lightgrey', color='lightgrey')+
  theme_void()#+
  #theme(panel.background = element_rect(color='black'))

traits.df<-read.csv('input_data/inputed_physio_traits_011923.csv')  %>%
  filter(species %in% mnm_spp) %>%
  dplyr::select(species, s.fossorial:Tmerge) %>%
  dplyr::select(-starts_with('Tforage'), -Tbask) %>%
  mutate(spF=factor(species, levels=rev(wtord))) %>%
  left_join(ecto_out %>% count(species) %>% rename(AOO=n), by='species')

wt.sum.df<-ect.df %>% group_by(species) %>% 
  summarize(medWT=median(WT_ptmin),
            WTmin2=quantile(WT_ptmin, .025),
            WTmin97=quantile(WT_ptmin, .975),
            nRows=n()) %>%
  mutate(spF=factor(species, levels=rev(wtord)))
library(cowplot)
plot_grid(
ect.df %>% 
  select(species, WT_ptmin, WT_min25, WT_min50) %>% 
  pivot_longer(-species) %>%
  group_by(species, name) %>% 
  summarize(medWT=median(value),
            WTmin2=quantile(value, .025),
            WTmin97=quantile(value, .975),
            nRows=n(),.groups='drop') %>%
  #left_join(rng_cent, by='species') %>%
  mutate(spF=factor(species, levels=rev(wtord)),
         nmF=factor(name, levels=rev(c('WT_ptmin','WT_min25','WT_min50')))) %>%
  ggplot()+
  geom_hline(yintercept=0, linetype='dashed', color='grey')+
  geom_linerange(aes(x=spF, ymin=WTmin2, ymax=WTmin97, group=nmF),
                 position=position_dodge(width=.4))+
  geom_point(aes(x=spF, y=medWT, shape=nmF),
             position=position_dodge(width=.4))+
  draw_plot(se_inset, x=1, y=-5, scale=10)+
  scale_y_continuous('Minimum warming tolerance', labels = tmplab)+
  scale_size('Quantification')+
  theme_classic()+coord_flip()+
  theme(axis.title.y=element_blank(),
        axis.text.y=element_text(face='italic'),
        #panel.grid.major.y = element_line(color='lightgrey'),
        legend.text = element_text(size=8),
        legend.title = element_text(size=8),
        legend.position='none'),
traits.df %>%
  select(spF, s.fossorial, CTmax, AOO, Tpref, Tmerge) %>%
  rename(fossorial=s.fossorial) %>%
  mutate(across(c(CTmax, AOO, Tpref, Tmerge), scale)) %>%
  pivot_longer(-spF) %>%
  mutate(traitF=factor(name, levels=c('AOO','fossorial','CTmax','Tpref','Tmerge'))) %>%
ggplot()+
  geom_tile(aes(x=traitF, y=spF, fill=value))+
  scale_x_discrete('Traits')+
  theme_classic()+
  theme(axis.title.y=element_blank(), axis.text.y=element_blank(),
        axis.line=element_blank(),  axis.ticks = element_blank(),
        legend.text = element_text(size=7),
        legend.title = element_text(size=7),
        axis.text.x = element_text(size=8, angle=45, hjust=1))+
  scale_fill_gradient2('Relative\ntrait\nvalue'),
rel_widths = c(.7,.3), align = 'h')
ggsave('results/Fig2_fancy_rank_thermalsensitivity_all.jpg', width=6, height=4)
```

Then, just to confirm the variation we see with our eyes, I run an ANOVA to see if there is a difference among different species.

```{r}
not.dif.sp<-NULL
spp.aov<-NULL
for(quant in c('WT_ptmin', 'WT_min25','WT_min50')){
  ect.df.a<-ect.df %>% rename(resp=all_of(quant))
  aov.min<-aov(resp~species, data=ect.df.a)
  aov.s<-anova(aov.min)
  thsd.min<-TukeyHSD(aov.min)$species
  not.dif.sp<-bind_rows(not.dif.sp,
                        thsd.min %>% as_tibble() %>% 
                          mutate(comp=rownames(thsd.min)) %>%
                          filter(`p adj` > 0.04) %>%
                          select(comp, everything()) %>%
                          mutate(quanti=quant))
  spp.aov<-bind_rows(spp.aov, aov.s %>% mutate(quanti=quant))
}
rm(ect.df.a)
```

```{r aov out, echo=F}
kable(spp.aov, caption='ANOVA results on daily minimum warming tolerance')
kable(not.dif.sp, caption='Species whose warming tolerance did not differ statistically')
```

I also conduct a one tailed Spearman's rho correlation to see if the ranked thermal sensitivity matches among all three quantifications of daily minimum warming tolerance. The null hypothesis is that there is not a positive correlation between the metrics.

```{r correlation among quants}
cor.test(ecto.summ$minWT, ecto.summ$WT25, method='spearman', alternative='g', exact=F)
cor.test(ecto.summ$minWT, ecto.summ$WT50, method='spearman', alternative='g', exact=F)
cor.test(ecto.summ$WT50, ecto.summ$WT25, method='spearman', alternative='g', exact=F)
```

Below is another way to visualize these different quantifications of daily minimum warming tolerances. This figure is in Appendix 2. 

```{r ridges plot, echo=F}
ect.df %>%
  select(species, WT_ptmin, WT_min25, WT_min50) %>%
  pivot_longer(-species) %>%
  mutate(spF=factor(species, levels=rev(wtord)),
          nmF=factor(name, levels=rev(c('WT_ptmin','WT_min25','WT_min50')))) %>%
  ggplot()+
  geom_density_ridges(aes(x=value, y=spF, fill=nmF), alpha=0.6)+
  scale_fill_grey(name='Point-based\nQuantification', labels=c('median','25th quantile','minimum'))+
  scale_x_continuous('Warming Tolerance', labels = tmplab, expand=c(0,0))+
  theme_classic()+
  theme(axis.title.y=element_blank(),
        axis.line.y.left = element_blank(),axis.ticks.y = element_blank(),
        axis.text.y=element_text(face='italic'),
        legend.text = element_text(size=8),
        legend.title = element_text(size=8))
ggsave('results/FigS2_3_wt_ridgelines.jpg', width=5, height=5)
```

Below is a peak at trait values for some of the most interesting species (in TPD's opinion).

```{r}
kable(traits.df %>% select(-spF) %>%
  filter(species %in% c('Gastrophryne carolinensis', 'Lithobates areolatus',
                        'Lithobates palustris', 'Lithobates clamitans')) %>% 
  pivot_longer(-species) %>% rename(Trait=name) %>% 
    pivot_wider(names_from=species, values_from = value),
  caption='Traits of our outlier species above. 1 in behavior traits means that behavior is allowed, 0 means it is not allowed (i.e., none of these species are allowed to climb)',
  digits=1)
```

## Cumulative warming tolerance across range

We can next count the proportion of the species range / sampled points that have negative (<0) or little (<2) warming tolerance. We use 2C as a heuristic cut off because of the common idea that temperatures will warm by 2 C. In the graph below, species whose curves are closer to the left are more thermally sensitive and have a higher proportion of their ranges with lower warming tolerance.

```{r cumdist plot, echo=F, fig.width=5, fig.height=2.5}
ect.df %>% 
  ggplot()+
    stat_ecdf(aes(x = WT_ptmin, color = pltsp), lwd=3, geom = "step", alpha=0.7)+
  stat_ecdf(aes(x = WT_ptmin, linetype=pltsp), geom = "step", alpha=0.9)+
  geom_vline(xintercept = c(0,2), color='grey')+
  scale_y_continuous('Cumulative Distribution')+
  scale_x_continuous(expression('Warming Tolerance ('*degree*C*')'),
                     breaks=seq(-12,12,2))+
  scale_color_manual('species', values=c(met.brewer('Renoir'), 'lightgrey'))+
                       #sppal[sppal$pltsp %in% ect.df$pltsp,]$col)+
  scale_linetype_manual('species', values=c(rep(c(1,2),6),1))+
                          #sppal[sppal$pltsp %in% ect.df$pltsp,]$shp)+
  theme_classic()+
  theme(legend.position='bottom', legend.direction = 'vertical',
        legend.margin = margin(t=-10,l=-30),
        legend.text = element_text(size=8,  face='italic'),
        axis.text = element_text(size=9),
        axis.title = element_text(size=9))+
    guides(color=guide_legend(nrow=3, size=2),
           linetype=guide_legend(nrow=3, size=2))
ggsave('results/Fig2_cumdist.jpg', width=5, height=3.5)
```

Quantifying this, we can see that some species exist very close to their thermal limits. 

```{r table of cumdist res, echo=F}
# calculate, of the occupied cells, the proportion that are below 0 
pdist<-ect.df %>%
  dplyr::select(species, WT_ptmin) %>%
  group_by(species) %>%
  summarize(n=n(),
            pdis_b2=round(ecdf(WT_ptmin)(2)*100,3),
            pdis_b0=round(ecdf(WT_ptmin)(0)*100,3),
            n_cellsb0=pdis_b0*n/100) %>% arrange(species)
kable(pdist %>%
    dplyr::select(-n),
  caption='Cumulative distribution of points where warming tolerance (estimated body temperature - CTmax) is below 0 and 2 degrees',
  col.names=c('Species', '% distribution <2', '% distribution < 0', 'n points <0'), digits=1)
```

```{r}
pdist %>% pull(pdis_b0) %>% summary()
pdist %>% filter(pdis_b0 == 0 | pdis_b0 == 100)
pdist %>% filter(pdis_b0 != 0) %>% arrange(pdis_b0)
cor(pdist$pdis_b0, pdist$pdis_b2, method='spearman')
pdist %>% select(species, starts_with('pdis')) %>%
  filter(pdis_b0 !=0) %>%
  mutate(across(where(is.numeric), rank)) %>%
  arrange(pdis_b0)
```

# Warming Tolerance across latitude

We then investigate if and how species warming tolerance changes across a latitudinal gradient. The latitudinal gradient represents a temperature gradient, along with other abiotic gradients, that could affect species warming tolerances. 

```{r figure 3 plots}
intra.res<-read.csv(PATH_intra_body) %>%
  filter(!is.na(loc)) %>% 
  left_join(read.csv(PATH_intra_loc), by='rowid')
lat.rng<-ect.df %>%
  group_by(species) %>%
  summarize(minlat=min(lat), maxlat=max(lat)) 
ndf<-data.frame(species=rep(unique(lat.rng$species), 30)) %>% 
  arrange(species) %>%
  mutate(lat=rep(seq.int(min(lat.rng$minlat), max(lat.rng$maxlat), length.out=30),
                 length(unique(lat.rng$species)))) %>%
  left_join(lat.rng, by='species') %>%
  filter( lat > minlat & lat < maxlat) 

mod.results <- NULL
for(quant in c('WT_ptmin','WT_min25','WT_min50')){
# to do: check model assumptions!
  ect.df1 <- ect.df %>% rename(QT=all_of(quant))
  test.mod<-lm(QT~lat:species + species + 0, data=ect.df1)
  (mod.sum<-summary(test.mod))

  m.est<-emtrends(test.mod, 'species', var='lat')
  mod.com.res<-pairs(m.est) %>% as_tibble()
  m.estT<-as_tibble(m.est)
  
  mod.results<-bind_rows(mod.results,
                         bind_rows(m.estT, mod.com.res) %>%
                           mutate(quant.var=quant))

  mod.pred<-cbind(ndf, predict(test.mod, newdata = ndf, interval='confidence')) %>%
    left_join(ect.df %>% distinct(species, pltsp), by='species')

  all.spp.p<-mod.pred %>% 
    ggplot()+
     geom_hline(yintercept=0, linetype='dashed')+
     geom_ribbon(aes(x=lat, ymin=lwr, ymax=upr, fill=pltsp), alpha=0.4)+
     geom_line(aes(x=lat, y=fit, color=pltsp), size=1.3)+
     geom_line(aes(x=lat, y=fit, linetype=pltsp), color='black', size=0.5)+
     scale_color_manual('species', 
                     values=c(met.brewer('Renoir')[1:6], 'lightgrey', met.brewer('Renoir')[7:12]),
                     aesthetics = c('fill','color')) +
    scale_linetype_manual('species', values=c(1,2,1,2,1,2,1,2,1,2,1,2,1))+
    scale_y_continuous(paste0(case_when(grepl('tmin', quant)~'Minimum',
                                        grepl('n25', quant)~'Hottest quarter',
                                        grepl('n50',quant)~'Median'), 
                              ' daily min. warming tolerance'), labels = tmplab)+
    scale_x_continuous('Latitude', labels=latlab, expand=c(0.01,0.01))+
    theme_classic()+
  theme(axis.text.x=element_text(size=8),
        legend.position = 'bottom', legend.direction = 'vertical',
        legend.text = element_text(size=8.5, face='italic'),
        legend.title=element_text(size=9),
        #legend.margin = margin(t=-4, l=-25),
        #plot.margin = margin(r=5, l=5)
        )+
  guides(fill=guide_legend(ncol=5),
         color=guide_legend(ncol=5))
intra.res1<-intra.res %>% rename(QT=all_of(quant))
intrawt.p<-intra.res1 %>%
  ggplot()+
  geom_smooth(data=ect.df1 %>% filter(species %in% unique(intra.res$species)),
              aes(x=lat, y=QT), color='black', method='lm', formula='y~x')+
  geom_smooth(aes(x=lat, y=QT, group=rn, color=species), 
              alpha=0.7, method='lm', formula='y~x')+
  facet_wrap(~species,)+
  scale_color_manual('species', 
                     values=c(met.brewer('Renoir')[c(2,1,8,11)]), #hand coded this
                     aesthetics = c('fill','color')) +
  scale_linetype_manual('species', values=c(2,1,1,2))+
  scale_y_continuous('Warming Tolerance', labels = tmplab)+
  scale_x_continuous('Latitude', labels=latlab, breaks=c(30,33,36), expand=c(0.01,0.01))+
  theme_classic()+
  theme(axis.text.x=element_text(size=8),
        axis.title.y=element_blank(),
        legend.position = 'none',
        strip.text=element_text(size=7.5, face='italic'),strip.background = element_blank(),
        panel.background = element_rect(fill=NA, color='lightgrey'))
lat.leg<-get_legend(all.spp.p)
plot_grid(plot_grid(all.spp.p+theme(legend.position='none'), intrawt.p, labels='AUTO'),
          lat.leg, rel_heights = c(.80,.20), ncol=1)
ggsave(paste0('results/Fig3_', quant,'_lat.jpg'), width=6.5, height=5)
}
```


```{r }
#c('WT_ptmin','WT_min25','WT_min50')
mod.min<-lm(WT_ptmin~lat:species + species + 0, data=ect.df)
mod.25<-lm(WT_min25~lat:species + species + 0, data=ect.df)
mod.50<-lm(WT_min50~lat:species + species + 0, data=ect.df)
plot(mod.min)
plot(mod.25)
plot(mod.50)
summary(mod.min)
summary(mod.25)
summary(mod.50)
deviance(mod.min)
extractAIC(mod.min)
```

```{r lin reg resul, echo=F}
lat.reg.coef<-mod.results %>% filter(!is.na(lat.trend)) %>%
  mutate(trend=paste0(round(lat.trend, 2), ' (', 
                      round(lower.CL, 2), ', ', round(upper.CL, 2), ')')) %>%
  select(species, quant.var, trend) %>%
  pivot_wider(names_from=quant.var, values_from = trend)
write.csv(lat.reg.coef, 'results/TableS3_latitude_reg_coeff.csv', row.names = F)
kable(lat.reg.coef, caption = 'Regression coefficients (95% confidence intervals) between daily minimum warming temperature and latitude', 
      col.names = c('Species','minimum WTd', '25th quartile WTd', 'mediant WTd'))
```

```{r}
mod.results %>% filter(is.na(lat.trend)) %>%
  select(contrast, p.value, quant.var) %>%
  filter(p.value > 0.05, quant.var=='WT_min25')

# lithobates areolatus not different from 10 species
grp.col<-data.frame(species=c(c('Acris blanchardi',  'Lithobates sylvaticus', 'Dryophytes versicolor', 'Gastrophryne carolinensis'),
                              c('Anaxyrus americanus','Lithobates palustris'),
                              c('Anaxyrus fowleri','Pseudacris crucifer')),
                    grp=c('ab','ab','ab','b','c','c','d','d'),
                    quant.var='WT_min25')


```


For quick comparison, the plot below plots the regression coefficients.

```{r regression coef plot, echo=F}
m.ord<-mod.results %>%  filter(quant.var=='WT_min25') %>%
  as_tibble() %>% arrange(lat.trend) %>% pull(species) 
mod.results %>% 
  filter(!is.na(lat.trend)) %>%
  left_join(grp.col, by=c('species','quant.var')) %>%
  mutate(spFac=factor(species, levels=m.ord))%>%
  ggplot()+
  geom_vline(xintercept=0, linetype='dashed', color='lightgrey')+
  geom_errorbarh(aes(y=spFac, xmin=lower.CL, xmax=upper.CL), height=0)+
  geom_point(aes(y=spFac, x=lat.trend, color=grp), size=3)+
  theme_classic()+
  facet_wrap(~quant.var, ncol=3)
```

Below, I use regressions to investigate how much warming tolerance varies when the intraspecific trait used to build the MNM varies. Below the differences among the different MNMs are displayed, with difference calculated as the lowest Warming Tolerance estimate - all other warming tolerance estimates. 

```{r intra dif plot, echo=F}
intra.traits<-read.csv('input_data/intraspecific_variation_subset_APPENDIX.csv') %>%
  filter(!is.na(traitSource)) %>%
  group_by(species) %>%
  mutate(rn=1:n()) %>%
  select(species, rn, Tpref, verbatimLocality, traitSource)
intra.traits

intra.res %>%
  left_join(intra.traits, by=c('species','rn','loc'='verbatimLocality')) %>%
  select(species, rowid, WT_ptmin, WT_min50, WT_min25, Tpref, rn, traitSource, loc) %>%
  group_by(species, rowid) %>%
  pivot_longer(cols=c('WT_ptmin','WT_min25','WT_min50')) %>%
  group_by(species, rowid, name) %>%
  arrange(value) %>%
  mutate(dif=lead(value, 1)-value, 
         sp_in=paste(species, rn),
         wt.f=factor(name, levels=c('WT_ptmin','WT_min25','WT_min50'),
                     labels=c('minimum','25th quartile','median'))) %>%
  filter(!is.na(dif)) %>%
  ggplot()+
  geom_boxplot(aes(x=as.character(Tpref), y=dif))+
  facet_grid(wt.f~species, scales='free')+ 
  scale_y_continuous('Deviation from lowest WT estimate')+
  scale_x_discrete('Tpref value')+
  theme_classic()+
  theme(strip.text.x = element_text(face='italic'))
ggsave('results/FigS2_6_intraspecific_trait_deviation.jpg',
       width=6.5, height=6)
```

```{r}
intra.mod.data<-intra.res %>%
  left_join(intra.traits, by=c('species','rn','loc'='verbatimLocality')) %>%
  select(species, rowid, WT_ptmin, WT_min50, WT_min25, Tpref, rn, traitSource, loc, lat) %>%
  mutate(sp_in=paste(species, rn))

all.intra.mod<-lm(WT_min25~lat:sp_in + sp_in +0, data=intra.mod.data)
intra.mod.out<-summary(all.intra.mod)
intra.traits %>%
  filter(!is.na(verbatimLocality)) %>%
  ungroup() %>%
  bind_cols(intra.mod.out$coefficients[grepl('lat', rownames(intra.mod.out$coefficients)),]) %>%
  select(-rn)  %>%
  mutate(across(where(is.numeric), ~round(.x, digits=3))) %>%
  write.csv('results/intraspecific_dif_coeffs.csv', row.names = F)
```

Below is a for loop to investigates species specific comparisons in differences among MNMs with different traits. 

```{r}
for(sp in unique(intra.mod.data$species)[1]){
  for(quant in 'WT_min25'){
    imdf<-intra.mod.data %>%
      rename(QT=all_of(quant)) %>%
      filter(species == sp)
    int.mod<-lm(QT~lat:sp_in + sp_in + 0, data=imdf)
summary(int.mod)
i.est<-emtrends(int.mod, 'sp_in', var='lat')
i.com.res<-pairs(i.est) %>% as_tibble() %>%
  mutate(quant) %>%
  select(quant, contrast, p.value, everything())
print(i.com.res %>% filter(p.value < 0.06))
}}
```

Below is code to run a phylogenetically informed linear regression. It needs to be run on the super computer and generally does not return different results than the model created above.

```{r phylo glm, eval=F}
library(ape);library(phyr)
taxa_table<-ect.df %>%
  distinct(species, pltsp) %>% 
  mutate(Tree.name.syn=case_when(
                             grepl('Dryophytes',species)~gsub('Dryophytes','Hyla', species),
                             species == 'Lithobates areolatus'~ 'Rana areolata',
                             species == 'Lithobates catesbeianus'~'Rana catesbeiana',
                             species == 'Lithobates sylvaticus'~ 'Rana sylvatica',
                             species == 'Lithobates sphenocephalus'~'Rana sphenocephala',
                             grepl('Lithobates',species)~gsub('Lithobates','Rana', species),
                             T~species),
         Tree.name=sub(' ','_',Tree.name.syn)) 

# import phylogeny from Jetz and Pyron 2018
tree_orig <- read.tree(file = "input_data/amph_shl_new_Consensus_7238.tre") 
tree.pruned <- keep.tip(phy = tree_orig, 
                        tip = taxa_table$Tree.name) 
ect.df2<-ect.df1 %>%
  left_join(taxa_table, by='species') %>%
  rename(spp=species, species = Tree.name) %>%
  mutate(latScale=scale(lat)) %>% ungroup() %>% sample_n(30000)
st.time<-Sys.time()
phlo.mod<-phyr::pglmm(QT~latScale + (1|species), 
            data=ect.df2, cov_ranef=list(species=tree.pruned))
end.time<-Sys.time()
end.time-st.time
phlo.mod
```

And then I output the data for use in the intraspecific comparison Rmd. 

```{r}
  mod.out.df<-mod.sum$coefficients[,-c(3,4)] %>% as_tibble() %>%
    bind_cols(parameter=rownames(mod.sum$coefficients)) %>%
    mutate(species = gsub('species', '',
                          gsub('lat:species','', parameter)),
         type=ifelse(grepl('lat:species',parameter), 'slope','intercept')) %>%
    select(-parameter) %>%
    pivot_wider(names_from=type, values_from=c(Estimate, `Std. Error`))
# output this data
ect.df %>% group_by(species) %>% 
   pivot_longer(cols=c('WT_min25','WT_ptmin', 'WT_min50')) %>%
  mutate(nm.F=factor(name, levels=c('WT_ptmin','WT_min25','WT_min50'),
                     labels=c('minimum','25th quartile','median'))) %>%
  group_by(species, nm.F) %>%
  summarize(medWT=median(value),
            WTmin2=quantile(value, .025),
            WTmin97=quantile(value, .975),
            nRows=n(), .groups='drop') %>%
  write.csv('results/pt_wtmin_sum.csv', row.names = F)

```

# quick investigation of longitudinal patterns of warming tolerance

```{r}
names(ect.df)
mod.lngmin<-lm(WT_ptmin~long:species + species + 0, data=ect.df)
summary(mod.lngmin)

mod.lng25<-lm(WT_min25~long:species + species + 0, data=ect.df)
summary(mod.lng25)
```

```{r}
ect.df %>%
  select(species, long, WT_ptmin, WT_min25) %>%
  pivot_longer(-c(species, long)) %>%
  ggplot()+
  geom_smooth(aes(x=long, y=value, color=species), method='lm') +
  facet_wrap(~name)
```
